### 2020-10-31 #2806 @whedon
[PRE REVIEW]: spaghetti: spatial network analysis in PySAL
**Submitting author:** @jGaboardi (<a href="http://orcid.org/0000-0002-4776-6826">James Gaboardi</a>)
**Repository:** <a href="https://github.com/pysal/spaghetti" target ="_blank">https://github.com/pysal/spaghetti</a>
**Version:** v1.5.3
