### 2020-10-31 #2806 @arfon
:wave: @usethedata - are you available to take on an additional paper currently? `@bstabler` would be the obvious editor to take this paper but he's currently unable to edit for JOSS.
