### 2020-11-07 #2806 @usethedata
@jGaboardi Thanks.  Nothing in particular at this time.  I have your list and a few others that have volunteered.  Unlike some other papers I've done, I don't think finding reviewers will be a huge problem.  Declines and no response aren't that unusual, and lots of people are in very challenging situations from time to time.

one more decline.  Additional request sent out this morning.
