+++ .github/workflows/unittests.yml
@@ -38,7 +38,7 @@
        - shell: bash -l {0}
          run: py.test -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
        - name: codecov (${{ matrix.os }}, ${{ matrix.environment-file }})
-         uses: codecov/codecov-action@v1.0.14
+         uses: codecov/codecov-action@v1.0.15
          with:
            token: ${{ secrets.CODECOV_TOKEN }}
            file: ./coverage.xml
