### 2020-11-17 7c0cc33
only check major release update for codecov
