### 2020-11-17 d197cd2
goanpeca -> conda-incubator for setup-minicoda
