+++ .github/workflows/unittests.yml
@@ -24,7 +24,7 @@
          #    os: windows-latest
      steps:
        - uses: actions/checkout@v2.3.4
-       - uses: goanpeca/setup-miniconda@v1
+       - uses: conda-incubator/setup-miniconda@v1
          with:
             miniconda-version: 'latest'
             auto-update-conda: true
