### 2020-11-17 237703b
v1 -> v2 conda-incubator/setup-miniconda
