+++ .github/workflows/unittests.yml
@@ -24,7 +24,7 @@
          #    os: windows-latest
      steps:
        - uses: actions/checkout@v2.3.4
-       - uses: conda-incubator/setup-miniconda@v1
+       - uses: conda-incubator/setup-miniconda@v2
          with:
             miniconda-version: 'latest'
             auto-update-conda: true
