+++ .github/workflows/unittests.yml
@@ -23,7 +23,7 @@
          #  - environment-file: ci/36-GIT-PLUS.yaml
          #    os: windows-latest
      steps:
-       - uses: actions/checkout@v2.3.4
+       - uses: actions/checkout@v2
        - uses: conda-incubator/setup-miniconda@v2
          with:
             miniconda-version: 'latest'
