+++ .github/workflows/unittests.yml
@@ -17,11 +17,11 @@
        matrix:
          os: [macos-latest, ubuntu-latest, windows-latest]
          environment-file: [ci/36-PYPI.yaml, ci/36-PYPI-PLUS.yaml, ci/36-GIT.yaml, ci/36-GIT-PLUS.yaml, ci/37-PYPI.yaml, ci/37-PYPI-PLUS.yaml, ci/37-GIT.yaml, ci/37-GIT-PLUS.yaml, ci/38-PYPI.yaml, ci/38-PYPI-PLUS.yaml, ci/38-GIT.yaml, ci/38-GIT-PLUS.yaml]
-         #exclude:
-         #  - environment-file: ci/36-PYPI-PLUS.yaml
-         #    os: windows-latest
-         #  - environment-file: ci/36-GIT-PLUS.yaml
-         #    os: windows-latest
+         exclude:
+           - environment-file: ci/36-PYPI-PLUS.yaml
+             os: windows-latest
+           - environment-file: ci/36-GIT-PLUS.yaml
+             os: windows-latest
      steps:
        - uses: actions/checkout@v2
        - uses: conda-incubator/setup-miniconda@v2
