+++ .github/workflows/release_and_publish.yml
@@ -15,9 +15,9 @@ jobs:
   build:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v2.3.4
+    - uses: actions/checkout@v2
     - name: Set up Python
-      uses: actions/setup-python@v2.1.4
+      uses: actions/setup-python@v2
       with:
         python-version: '3.x'
     - name: Install dependencies
@@ -37,7 +37,7 @@ jobs:
         TEXT: ${{ steps.changetxt.outputs.text }}
     - name: Create Release
       id: create_release
-      uses: actions/create-release@v1.1.4
+      uses: actions/create-release@v1
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # token is provided by GHA, DO NOT create
       with:
@@ -50,7 +50,7 @@ jobs:
       run: |
         export PKG=$(ls dist/)
         set -- $PKG
-        echo "::set-env name=whl_path::$1"
+        echo "whl_path=1" >> $GITHUB_ENV
     - name: Upload Release Asset
       id: upload-release-asset 
       uses: actions/upload-release-asset@v1
