### 2020-11-22 50fe898
Revert "Python 3.9 testing for CI"
