+++ ci/39-GIT.yaml
@@ -1,21 +1,0 @@
-name: test
-channels:
-  - conda-forge
-dependencies:
-  - python=3.9
-  # required
-  - esda
-  - numpy
-  - pandas>=1.0
-  - rtree
-  - scipy>=1.0
-  - watermark
-  # testing/formatting
-  - black
-  - codecov
-  - pytest
-  - pytest-cov
-  # required
-  - pip
-  - pip:
-    - git+https://github.com/pysal/libpysal.git@master
