+++ ci/39-GIT-PLUS.yaml
@@ -1,23 +1,0 @@
-name: test
-channels:
-  - conda-forge
-dependencies:
-  - python=3.9
-  # required
-  - esda
-  - numpy
-  - pandas>=1.0
-  - rtree
-  - scipy>=1.0
-  - watermark
-  # testing/formatting
-  - black
-  - codecov
-  - pytest
-  - pytest-cov
-  # optional
-  - geopandas>=0.7.0
-  # required
-  - pip
-  - pip:
-    - git+https://github.com/pysal/libpysal.git@master
