+++ .github/workflows/unittests.yml
@@ -36,7 +36,7 @@
        - shell: bash -l {0}
          run: conda list
        - shell: bash -l {0}
-         run: py.test -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
+         run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
        - name: codecov (${{ matrix.os }}, ${{ matrix.environment-file }})
          uses: codecov/codecov-action@v1
          with:
