+++ .github/dependabot.yml
@@ -12,6 +12,13 @@ updates:
     reviewers:
       - "jGaboardi"
 
+  - package-ecosystem: "gitsubmodule"
+    directory: "/"
+    schedule:
+      interval: "daily"
+    reviewers:
+      - "jGaboardi"
+
   - package-ecosystem: "pip"
     directory: "/"
     schedule:
