### 2020-12-25 175a973
add default shell to unittests
