### 2020-12-26 09e00cc
update setup-miniconda in unittests
