+++ .github/workflows/unittests.yml
@@ -32,7 +32,7 @@
            CACHE_NUMBER: 0
          with:
            path: ~/conda_pkgs_dir
-           key: ${{ runner.os }}-conda-${{ env.CACHE_NUMBER }}-${{ hashFiles(matrix.environment-file) }}
+           key: ${{ matrix.os }}-conda-${{ env.CACHE_NUMBER }}-${{ hashFiles(matrix.environment-file) }}
        - uses: conda-incubator/setup-miniconda@v2
          with:
             miniconda-version: 'latest'
