### 2020-12-27 238ef36
remove sphinx_gallery from reqs
