+++ requirements_docs.txt
@@ -3,5 +3,3 @@ numpydoc
 sphinx>=1.4.3
 sphinxcontrib-bibtex
 sphinx_bootstrap_theme
-sphinx_gallery
-
