+++ docs/index.rst
@@ -1,0 +1,137 @@
+.. documentation master file
+
+
+.. raw:: html
+
+    <img 
+        src="_static/images/spaghetti_nav_logo.svg" 
+        class="img-responsive center-block" 
+        alt="spaghetti logo" 
+        width="470" 
+        height="200"
+    >
+
+
+`spaghetti`
+===========
+
+**SPA**\ tial **G**\ rap\ **H**\ s: n\ **ET**\ works, **T**\ opology, & **I**\ nference
+---------------------------------------------------------------------------------------
+
+`Spaghetti` is an open-source Python library for the analysis of network-based
+spatial data. Originating from the `network` module in `PySAL (Python Spatial 
+Analysis Library) <http://pysal.org>`_, it is under active development for the 
+inclusion of newly proposed methods for building graph-theoretic networks and 
+the analysis of network events. An installation guide, API reference, 
+and usage tutorials are provided here through the links above.
+
+
+.. raw:: html
+
+    <div class="container-fluid">
+      <div class="row equal-width">
+        <div class="col-sm-.5 col-xs-hidden">
+        </div>
+        <div class="col-md-6 col-xs-15">
+            <a 
+                href="https://pysal.org/spaghetti/notebooks/network-segmentation.html" 
+                class="thumbnail"
+            >
+                <img 
+                    src="_static/images/crime_counts.png" 
+                    class="img-responsive center-block"
+                >
+                <div class="caption text-center">
+                <h6>Network Representations</h6>
+                </div>
+            </a>
+        </div>
+        <div class="col-sm-6 col-xs-15">
+            <a 
+                href="https://pysal.org/spaghetti/notebooks/network-spatial-dependence.html" 
+                class="thumbnail"
+            >
+                <img 
+                    src="_static/images/network_k.png" 
+                    class="img-responsive center-block"
+                >
+                <div class="caption text-center">
+                <h6>Network Spatial Dependence</h6>
+                </div>
+            </a>
+        </div>
+        </div>
+        <div class="col-sm-.5 col-xs-hidden">
+        </div>
+      </div>
+    </div>
+
+
+Development
+-----------
+
+Development of `spaghetti` is hosted on GitHub_.
+
+
+Citing `spaghetti`
+------------------
+
+If you use PySAL-spaghetti in a scientific publication, we would appreciate using the following BibTeX citation::
+
+  @misc{Gaboardi2018,
+    author   = {Gaboardi, James D. and Laura, Jay and Rey, Sergio and
+                Wolf, Levi John and Folch, David C. and Kang, Wei and 
+                Stephens, Philip and Schmidt, Charles},
+    month    = {oct},
+    year     = {2018},
+    title    = {pysal/spaghetti},
+    url      = {https://github.com/pysal/spaghetti},
+    doi      = {10.5281/zenodo.1343650},
+    keywords = {graph-theory,network-analysis,python,spatial-networks,topology}
+  }
+
+
+Funding
+-------
+
+This project is/was partially funded through:
+
+.. figure:: _static/images/ardc_logo.png
+    :target: https://atlantardc.wordpress.com
+    :width: 150
+    :align: left
+
+    The Atlanta Research Data Center: `A Polygon-Based Approach to Spatial Network Allocation <https://atlantardc.files.wordpress.com/2018/05/ardc-newsletter_2018_2.pdf>`_
+
+.. figure:: _static/images/nsf_logo.png
+    :target: https://www.nsf.gov/index.jsp
+    :width: 100
+    :align: left
+
[…]
