+++ .github/workflows/build_docs.yml
@@ -1,0 +1,61 @@
+
+ name: Build Docs
+ on:
+   push:
+     branches:
+     - main
+ jobs:
+   docs:
+     name: Build & Push Docs
+     runs-on: ${{ matrix.os }}
+     timeout-minutes: 90
+     strategy:
+       matrix:
+         os: ['ubuntu-latest']
+         environment-file: [ci/39.yml]
+         experimental: [false]
+     defaults:
+       run:
+         shell: bash -l {0}
+     steps:
+       - uses: actions/checkout@v2
+       - uses: actions/cache@v2
+         env:
+           CACHE_NUMBER: 0
+         with:
+           path: ~/conda_pkgs_dir
+           key: ${{ matrix.os }}-conda-${{ env.CACHE_NUMBER }}-${{ hashFiles(matrix.environment-file) }}
+       - uses: conda-incubator/setup-miniconda@v2
+         with:
+            miniconda-version: 'latest'
+            mamba-version: '*'
+            channels: conda-forge
+            channel-priority: true
+            auto-update-conda: false
+            auto-activate-base: false
+            environment-file: ${{ matrix.environment-file }}
+            activate-environment: test
+            use-only-tar-bz2: true
+       - run: mamba info --all
+       - run: mamba list
+       - run: mamba config --show-sources
+       - run: mamba config --show
+       - name: Make Docs
+         run: cd docs; make html
+       - name: Commit Docs
+         run: |
+           git clone https://github.com/ammaraskar/sphinx-action-test.git --branch gh-pages --single-branch gh-pages
+           cp -r docs/_build/html/* gh-pages/
+           cd gh-pages
+           git config --local user.email "action@github.com"
+           git config --local user.name "GitHub Action"
+           git add .
+           git commit -m "Update documentation" -a || true
+           # The above command will fail if no changes were present,
+           # so we ignore the return code.
+       - uses: ad-m/github-push-action@master
+         with:
+            branch: gh-pages
+            directory: gh-pages
+            github_token: ${{ secrets.GITHUB_TOKEN }}
+            force: true
