+++ docsrc/conf.py
@@ -1,308 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# SPAGHETTI documentation build configuration file, based on `giddy`. This file is
-# execfile()d with the current directory set to its containing dir. Note that not all
-# possible configuration values are present in this autogenerated file. All configuration
-# values have a default; values that are commented out serve to show the default. If
-# extensions (or modules to document with autodoc) are in another directory, add these
-# directories to sys.path here. If the directory is relative to the documentation root,
-# use os.path.abspath to make it absolute, like shown here.
-
-import sys, os
-import sphinx_bootstrap_theme
-
-
-sys.path.insert(0, os.path.abspath("../../"))
-
-# import your package to obtain the version info to display on the docs website
-import spaghetti
-
-
-# -- General configuration ------------------------------------------------
-
-# If your documentation needs a minimal Sphinx version, state it here.
-#
-# needs_sphinx = '1.0'
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
-extensions = [  #'sphinx_gallery.gen_gallery',
-    "sphinx.ext.autodoc",
-    "sphinx.ext.autosummary",
-    "sphinx.ext.viewcode",
-    "sphinxcontrib.bibtex",
-    "sphinx.ext.mathjax",
-    "sphinx.ext.doctest",
-    "sphinx.ext.intersphinx",
-    "numpydoc",
-    "nbsphinx",
-]
-
-
-# Add any paths that contain templates here, relative to this directory.
-templates_path = ["_templates"]
-
-# The suffix(es) of source filenames.
-# You can specify multiple suffix as a list of string:
-#
-# source_suffix = ['.rst', '.md']
-source_suffix = ".rst"
-
-# The master toctree document.
-master_doc = "index"
-
-# General information about the project.
-project = "spaghetti"
-copyright = "2018, pysal developers"
-author = "pysal developers"
-
-# The version info for the project you're documenting, acts as replacement for
-# |version| and |release|, also used in various other places throughout the
-# built documents.
-#
-# The full version.
-version = spaghetti.__version__
-release = spaghetti.__version__
-
-# The language for content autogenerated by Sphinx. Refer to documentation
-# for a list of supported languages.
-#
-# This is also used if you do content translation via gettext catalogs.
-# Usually you set "language" from the command line for these cases.
-language = None
-
-# List of patterns, relative to source directory, that match files and
-# directories to ignore when looking for source files.
-# This patterns also effect to html_static_path and html_extra_path
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "tests/*"]
-
-# The name of the Pygments (syntax highlighting) style to use.
-pygments_style = "sphinx"
-
-# If true, `todo` and `todoList` produce output, else they produce nothing.
-todo_include_todos = False
-
-# -- Options for HTML output ----------------------------------------------
-
-# The theme to use for HTML and HTML Help pages.  See the documentation for
-# a list of builtin themes.
-#
-# html_theme = "alabaster"
-html_theme = "bootstrap"
-html_theme_path = sphinx_bootstrap_theme.get_html_theme_path()
-html_title = "%s v%s Manual" % (project, version)
-
-# (Optional) Logo of your package. Should be small enough to fit the navbar (ideally 24x24).
-# Path should be relative to the ``_static`` files directory.
-# html_logo = "_static/images/package_logo.jpg"
[…]
