### 2020-12-28 666c63e
mamba->conda for config @knaaptime
