+++ .github/workflows/build_docs.yml
@@ -38,8 +38,8 @@
             use-only-tar-bz2: true
        - run: mamba info --all
        - run: mamba list
-       - run: mamba config --show-sources
-       - run: mamba config --show
+       - run: conda config --show-sources
+       - run: conda config --show
        - name: Make Docs
          run: cd docs; make html
        - name: Commit Docs
