### 2020-12-29 d491da3
add packages for build_docs.yml
