+++ ci/39-GIT-PLUS.yaml
@@ -17,6 +17,12 @@ dependencies:
   - pytest-cov
   # optional
   - geopandas>=0.7.0
+  # for docs build action (this env only)
+  - nbsphinx
+  - numpydoc
+  - sphinx>=1.4.3
+  - sphinxcontrib-bibtex
+  - sphinx_bootstrap_theme
   # required
   - pip
   - pip:
