### 2020-12-29 0567652
update sys.path
update sys.path due to [this error](https://github.com/pysal/spaghetti/runs/1622831881?check_suite_focus=true#step:9:19).
