### 2020-12-29 df5e1cb
pin sphinxcontrib-bibtex version
pin sphinxcontrib-bibtex to `<2.0.0` due to [this issue](https://github.com/executablebooks/jupyter-book/issues/1137#issuecomment-743994864)
