+++ ci/39-GIT-PLUS.yaml
@@ -21,7 +21,7 @@ dependencies:
   - nbsphinx
   - numpydoc
   - sphinx>=1.4.3
-  - sphinxcontrib-bibtex
+  - sphinxcontrib-bibtex<2.0.0
   - sphinx_bootstrap_theme
   # required
   - pip
