+++ docs/conf.py
@@ -265,8 +265,8 @@ intersphinx_mapping = {
 
 # This is processed by Jinja2 and inserted before each notebook
 nbsphinx_prolog = r"""
-{% set docname = env.doc2path(env.docname, base='docsrc/') %}
-{% set fullpath = env.doc2path(env.docname, base='tree/master/docsrc/') %}
+{% set docname = env.doc2path(env.docname, base='docs/') %}
+{% set fullpath = env.doc2path(env.docname, base='tree/main/docs/') %}
 
 .. only:: html
 
@@ -277,7 +277,7 @@ nbsphinx_prolog = r"""
 
         This page was generated from `{{ docname }}`__.
         Interactive online version:
-        :raw-html:`<a href="https://mybinder.org/v2/gh/pysal/spaghetti/master?filepath={{ docname }}"><img alt="Binder badge" src="https://mybinder.org/badge_logo.svg" style="vertical-align:text-bottom"></a>`
+        :raw-html:`<a href="https://mybinder.org/v2/gh/pysal/spaghetti/main?filepath={{ docname }}"><img alt="Binder badge" src="https://mybinder.org/badge_logo.svg" style="vertical-align:text-bottom"></a>`
 
     __ https://github.com/pysal/spaghetti/{{ fullpath }}
 
