### 2020-12-29 1376b62
clean out docs dir and rerun notebooks
