+++ docs/_sources/api.rst.txt
@@ -1,40 +1,0 @@
-.. _api_ref:
-
-.. currentmodule:: spaghetti
-
-API reference
-=============
-
-.. _network_api:
-
-
-Network feature instantiation
------------------------------
-
-.. autosummary::
-   :toctree: generated/
-    
-    spaghetti.Network
-    spaghetti.PointPattern
-
-
-Network feature extraction and creation
----------------------------------------
-
-.. autosummary::
-   :toctree: generated/
-    
-    spaghetti.extract_component
-    spaghetti.spanning_tree
-    spaghetti.element_as_gdf
-    spaghetti.regular_lattice
-
-
-Save and load a network
------------------------
-
-.. autosummary::
-   :toctree: generated/
-    
-    spaghetti.Network.savenetwork
-    spaghetti.Network.loadnetwork
