+++ docs/_sources/tutorials.rst.txt
@@ -1,61 +1,0 @@
-.. tutorials
-
-
-Tutorials
-=========
-
-.. toctree::
-    :maxdepth: 1
-    :caption: Basic functionality:
-
-    notebooks/quickstart.ipynb
-    notebooks/pointpattern-attributes.ipynb
-    notebooks/connected-components.ipynb
-    notebooks/shortest-path-visualization.ipynb
-    
-.. toctree::
-    :maxdepth: 1
-    :caption: Advanced functionality and known limitations:
-    
-    notebooks/network-segmentation.ipynb
-    notebooks/network-spatial-dependence.ipynb
-    notebooks/network-spatial-weights.ipynb
-    notebooks/spanning-trees.ipynb
-    notebooks/caveats.ipynb
-
-.. toctree::
-    :maxdepth: 1
-    :caption: Integrations & applications:
-    
-    notebooks/network-spatial-autocorrelation.ipynb
-    notebooks/facility-location.ipynb
-    notebooks/transportation-problem.ipynb
-    notebooks/tsp.ipynb
-
-Citations:
-  * Advanced functionality:
-  
-    * Spatial Dependence:
-      :cite:`Ripley1976`, :cite:`Ripley1977`, :cite:`doi:10.1002/9780470549094.ch5`, :cite:`doi:10.1111/j.1538-4632.2001.tb00448.x`, :cite:`doi:10.1002/9781119967101.ch6`
-    
-    * Spatial Weights:
-      :cite:`pysal2007`, :cite:`AnselinRey2014`, :cite:`rey_open_2015`
-    
-    * Spanning Trees:
-      :cite:`GrahamHell_1985`, :cite:`AhujaRavindraK`, :cite:`doi:10.1002/9781119967101.ch3`
-  
-  * Integrations & applications:
-    
-    * Spatial Autocorrelation:
-      :cite:`Moran1950`, :cite:`Anselin1995`, :cite:`Getis1992`, :cite:`esda:_2019`, :cite:`splot:package:_2020`, :cite:`Lumnitz2020`
-    
-    * Facility Location:
-      :cite:`Hakimi1964`, :cite:`ReVelle1970`, :cite:`Toregas1971`, :cite:`Toregas1972`, :cite:`Church1974`, :cite:`ReVelle2005`
-    
-    * The Transportation Problem:
-      :cite:`Hitchcock1941`, :cite:`Koopmans1949`, :cite:`Phillips1981a`, :cite:`Gass2005`, :cite:`Church2009`, :cite:`daskin2013`
-    
-    * The Traveling Salesperson Problem:
-      :cite:`Dantzig1954`, :cite:`Flood1956`, :cite:`Miller1960`, :cite:`Cummings2000`, :cite:`Miller2001`, :cite:`Bektas2014`
-
-
