+++ docs/_sources/generated/spaghetti.Network.rst.txt
@@ -1,40 +1,0 @@
-spaghetti.Network
-=================
-
-.. currentmodule:: spaghetti
-
-.. autoclass:: Network
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~Network.GlobalAutoK
-      ~Network.__init__
-      ~Network.allneighbordistances
-      ~Network.compute_distance_to_vertices
-      ~Network.compute_snap_dist
-      ~Network.contiguityweights
-      ~Network.count_per_link
-      ~Network.distancebandweights
-      ~Network.enum_links_vertex
-      ~Network.extractgraph
-      ~Network.full_distance_matrix
-      ~Network.identify_components
-      ~Network.loadnetwork
-      ~Network.nearestneighbordistances
-      ~Network.savenetwork
-      ~Network.shortest_paths
-      ~Network.simulate_observations
-      ~Network.snapobservations
-      ~Network.split_arcs
-   
-   
-
-   
-   
-   
