### 2020-12-30 a0596e0
update links to notebooks
