+++ docs/conf.py
@@ -265,8 +265,8 @@ intersphinx_mapping = {
 
 # This is processed by Jinja2 and inserted before each notebook
 nbsphinx_prolog = r"""
-{% set docname = env.doc2path(env.docname, base='docs/') %}
-{% set fullpath = env.doc2path(env.docname, base='tree/main/docs/') %}
+{% set docname = env.doc2path(env.docname, base='notebooks/') %}
+{% set fullpath = env.doc2path(env.docname, base='tree/main/notebooks/') %}
 
 .. only:: html
 
