### 2020-12-30 fadd1ec
try non-duplicate notebook dir
