+++ docs/conf.py
@@ -266,7 +266,7 @@ intersphinx_mapping = {
 # This is processed by Jinja2 and inserted before each notebook
 nbsphinx_prolog = r"""
 {% set docname = env.doc2path(env.docname, base='') %}
-{% set fullpath = env.doc2path(env.docname, base='tree/main/notebooks/') %}
+{% set fullpath = env.doc2path(env.docname, base='tree/main/') %}
 
 .. only:: html
 
