### 2021-01-04 #2826 @jGaboardi
> Or should I wait for the amendments that @jGaboardi is planning to include?

The release is a scheduled one to coincide with the biannual PySAL federation release. It will not include any change to the actual code-base, just admin/maintenance adjustments for testing, docs, etc. I will get that out today, which will be `v1.5.4`.
