### 2021-01-04 2c81619
correct GITHUB_ENV in release workflow
