+++ .github/workflows/release_and_publish.yml
@@ -50,7 +50,7 @@ jobs:
       run: |
         export PKG=$(ls dist/)
         set -- $PKG
-        echo "whl_path=1" >> $GITHUB_ENV
+        echo "name=$1" >> $GITHUB_ENV
     - name: Upload Release Asset
       id: upload-release-asset 
       uses: actions/upload-release-asset@v1
