+++ spaghetti/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.5.4"
+__version__ = "1.5.5"
 
 """
 # `spaghetti` --- Spatial Graphs: Networks, Topology, & Inference
