### 2021-01-11 pysal/spaghetti#574 @martibosch
[JOSS review] Manuscript summary
The "Summary" section of the manuscript is too short and does not provide enough information about the purpose of the library - other than telling the reader that it is a library about spatial networks. I would suggest extending the sentence "To this end, spaghetti (spatial graphs: networks, topology, & inference), a sub-module in the wider PySAL ecosystem (see Figure 1), was developed to address network-centric research questions with a strong focus on spatial analysis." with a very brief overview of the main functionalities of the library. I believe that by reading the summary, a potential user should be able to assess whether the tool is of its interest.

https://github.com/openjournals/joss-reviews/issues/2826
