### 2021-01-11 pysal/spaghetti#575 @martibosch
[JOSS review] Manuscript statement of need
The "Statement of need" section of the manuscript does not actually feature a statement of need but rather a short review of related tools. I believe that a proper statement of need could be added after the first (and only) paragraph of the section. Such a statement should explain which aspects of spaghetti represent an improvement over the reviewed tools as well as who the target audience is. In line with the comment above, the authors might actually consider collapsing this section with the summary (it is just a suggestion, but I understand if the authors prefer having two separate sections - see also #574).

https://github.com/openjournals/joss-reviews/issues/2826
