### 2021-01-11 pysal/spaghetti#576 @martibosch
[JOSS review] Manuscript Figure 1 (PySAL logo)
I think that Figure 1 of the manuscript (i.e., the PySAL logo) is not needed and should hence be deleted - especially considering that JOSS papers tend to be very brief.
https://github.com/openjournals/joss-reviews/issues/2826
