### 2021-01-11 pysal/spaghetti#577 @martibosch
[JOSS review] Add more information in the tutorial notebooks
The notebooks from the tutorial provide a very easy way (especially with the Binder badges) for the users to interactively follow the tutorial, nonetheless they could have more text explaining the functionalities that they showcase. This might not be much of an issue for and advanced user who can infer many information from the context and browse the API documentation and source code of the classes and methods used in the notebooks, yet I think that with the current notebooks, many users will not know what to do other than hitting enter to execute each cell sequentially. I believe that the user guides of momepy (https://docs.momepy.org/en/stable/user_guide/intro.html) or OSMnx (https://github.com/gboeing/osmnx-examples) are great examples of a thorough tutorials.
https://github.com/openjournals/joss-reviews/issues/2826
