### 2021-01-11 pysal/spaghetti#578 @martibosch
[JOSS review] Improve the quickstart notebook
In the line of #577, the quickstart notebook could be a bit more thorough in showcasing the main functionalities of spaghetti. For instance, it first shows the user how to instantiate a `spaghetti.Network` object, it could mention other ways of instantiating such an object, e.g., from `libpysal.cg.Chain` as done in the connected components notebook or as a regular lattice as done in the shortest path visualization notebook. This could be done in the form of a markdown cell with a couple of brief sentences and the respective links to the ohter noteooks of the tutorial as well as the corresponding API documentation.
https://github.com/openjournals/joss-reviews/issues/2826
