### 2021-01-11 pysal/spaghetti#579 @martibosch
[JOSS review] Incomplete API documentation site
The [API documentation](https://pysal.org/spaghetti/api.html) site is incomplete, e.g., missing `spaghetti.network.Network.snapobservations`, `spaghetti.network.Network.split_arcs` and many other relevant methods (which are in fact well documented in their respective docstrings).

https://github.com/openjournals/joss-reviews/issues/2826
