### 2021-01-11 pysal/spaghetti#580 @martibosch
[JOSS review] Network class, key (unique) functionalities of spaghetti, relationship to NetworkX (or the like).
A central point that I would like to raise is why does spaghetti define its own network class when there exist plenty of Python tools to that purpose (most notably NetworkX). For instance, basic network methods featured in spaghetti such as the shortest paths, connected components, spanning trees and the like are already implemented in existing tools. Isn't re-implementing these features in spaghetti a bit "re-inventing the wheel"?

On the other hand, I have the impression many of the functionalities that seem unique to spaghetti, such as snapping points, splitting arcs, network weights and the like could also be implemented by extending an existing network class (again, e.g., NetworkX).

Of course, I am sure that the authors have already considered this - I would just like to raise this issue in the reviewing. Hence, I would like to ask them why a custom class has been created for an object as basic as a network (for which so many Python tools exist already).
https://github.com/openjournals/joss-reviews/issues/2826
