### 2021-01-11 pysal/spaghetti#580 @martibosch
I am sorry if I am repeating myself, but: related to this issue, considering that there exist several libraries to model networks in Python, I believe that the unique functionalities of spaghetti (those that are not implemented in other Python libraries such as NetworkX) should be listed more clearly in the README, tutorials, API documentation and the manuscript (see #575).

More precisely, the following questions should be answered clearly and concisely: why do we need another library to model networks in Python? which functionalities does spaghetti include that the other libraries don't?
