### 2021-01-11 pysal/spaghetti#581 @martibosch
[JOSS review] Methods to assist common tasks (especially plotting)
The notebooks of the tutorial show how spaghetti can be used to assist a variety of spatial analysis tasks related to spatial networks. Nevertheless, to that end, the notebooks require a substantial amount of ad-hoc code, especially when it comes to plotting tasks. A key example concerns the simplest network visualization task, which seemingly requires using the `spaghetti.element_as_gdf` method to obtain geo-data frames for the vertices and arcs and then using the corresponding geopandas plotting methods. Wouldn't it be possible to create method(s) (maybe in the `utils.py` or as a separate, e.g., `plotting.py` module) to automatize these common tasks and offer a nicer API to the users so that they do not need to write too much repetitive and ad-hoc code? Such a method could include arguments to automatically add labels and legends for the nodes and arcs, and to assist further customization for the plots. I am thinking about something like the [`plot_graph` method](https://osmnx.readthedocs.io/en/stable/osmnx.html#osmnx.plot.plot_graph) of OSMnx or more broadly its [`plot` module](https://osmnx.readthedocs.io/en/stable/osmnx.html#module-osmnx.plot).

After browsing the tutorial notebooks, the case for a plotting method(s) appears very relevant, yet the same remark might apply to further recurrent tasks (other than plotting) for which spaghetti has been design to assist.

https://github.com/openjournals/joss-reviews/issues/2826
