### 2021-01-11 pysal/spaghetti#582 @martibosch
[JOSS review] Support reading/writing networks from different formats, integration with NetworkX/OSMnx
The [`in_data` argument of spaghetti.Network](https://pysal.org/spaghetti/generated/spaghetti.Network.html#spaghetti.Network) lists six possible input types but only describes four of them.
The notebooks of the tutorials use example data from PySAL in the form of a shapefile and iterables of `libpysal.cg.Chain`, nonetheless, the users might have their network data in other (more) common data formats (see, e.g., [the formats supported by NetworkX](https://networkx.org/documentation/stable/reference/readwrite/index.html)). Would it be possible to instantiate `spaghetti.Network` objects from NetworkX graph objects so that spaghetti delegates the I/O part to NetworkX and focuses on the spatial analysis functionalities (see also #580). I know that there are other Python librares to deal with NetworkX, but since I suppose that spaghetti is mostly concerned with street networks, supporting NetworkX objects has the major advantage of supporting in turn OSMnx, which makes it very easy to stream street networks from OSM directly into Python objects.

Again, the developers might have good reasons (or other priorities) to not work with NetworkX objects, the point of this issue is to ensure that the matter has been considered.

https://github.com/openjournals/joss-reviews/issues/2826
