### 2021-01-11 pysal/spaghetti#583 @martibosch
[JOSS review] Manuscript "Current functionality"
The "Current functionality" section of the manuscript only shows one of the functionalities of spaghetti, namely the instantiation of a regular lattice network. The rest of the code snippet is more about geopandas and matplotlib. In line with #575 and #580, I think that the section should make a better case for the need of spaghetti in the Python stack (perhaps snapping points?).

https://github.com/openjournals/joss-reviews/issues/2826
