### 2021-01-11 #2826 @martibosch
This would be all from my side. Looking forward to hear back from the authors. Let me clear out again that my intention is not to question the criteria of the authors (since they probably know better than me why they have made such choices), but rather ensure that the raised issues have been considered.

Best,
Martí
