### 2021-01-12 pysal/spaghetti#579 @jGaboardi
@martibosch the API docs for `spaghetti.Network.snapobservations` are [here](https://pysal.org/spaghetti/generated/spaghetti.Network.html#spaghetti.Network.snapobservations) and `spaghetti.network.Network.split_arcs` are [here](https://pysal.org/spaghetti/generated/spaghetti.Network.html#spaghetti.Network.split_arcs). Which were the others that were missing?
