### 2021-01-12 pysal/spaghetti#579 @martibosch
my mistake 😓 I did not find them listed in the [API documentation](https://pysal.org/spaghetti/api.html) page but this is because they are actually methods of the Network class (which does appear in the page). Closing 💪🏻
