### 2021-01-12 #2826 @usethedata
@martibosch -- Thanks.
@jGaboardi -- My apologies for being missing in action on this.  The last several weeks have been a bit challenging.  I'll set the version on the software for the paper as one of the last things in the review process, so there aren't issues there.  I'll work on the second reviewer shortly.
