+++ paper/paper.bib
@@ -34,7 +34,7 @@
 }
 
 @misc{gaboardi2018,
-    author    = {Gaboardi, James D. and Laura, Jay and Rey, Sergio and 
+    author = {Gaboardi, James D. and Laura, Jay and Rey, Sergio and 
                  Wolf, Levi John and Folch, David C. and Kang, Wei and 
                  Stephens, Philip and Schmidt, Charles},
     month     = {oct},
@@ -46,7 +46,7 @@
 }
 
 @article{okabe2006a,
-    author = {Okabe, Atsuyuki and Okunuki, Keiichi and Shiode,     Shino},
+    author = {Okabe, Atsuyuki and Okunuki, Keiichi and Shiode, Shino},
     doi = {10.1111/j.0016-7363.2005.00674.x},
     isbn = {1538-4632},
     issn = {0016-7363},
@@ -84,7 +84,7 @@
     journal = {Computers, Environment and Urban Systems},
     pages = {126--139},
     publisher = {Elsevier Ltd},
-    title = {{OSMnx: New Nethods for Acquiring, Constructing, Analyzing, and Visualizing Complex Street Networks}},
+    title = {{OSMnx: New Methods for Acquiring, Constructing, Analyzing, and Visualizing Complex Street Networks}},
     volume = {65},
     year = {2017}
 }
@@ -205,4 +205,18 @@
   author={Foti, Fletcher and Waddell, Paul and Luxen, Dennis},
   booktitle={Proceedings of the 4th TRB Conference on Innovations in Travel Modeling. Transportation Research Board},
   year={2012}
+}
+
+@inbook{osullivan_unwin2010.ch5,
+    author = {O'Sullivan, D. and Unwin, D. J.},
+    publisher = {John Wiley \& Sons, Ltd},
+    isbn = {9780470549094},
+    title = {Point Pattern Analysis},
+    booktitle = {Geographic Information Analysis},
+    chapter = {5},
+    pages = {121-156},
+    doi = {10.1002/9780470549094.ch5},
+    year = {2010},
+    keywords = {point separation, standard deviation, observed VMR, estimation methods, computationally intensive},
+    abstract = {Summary This chapter contains sections titled: Introduction Describing a Point Pattern Assessing Point Patterns Statistically Monte Carlo Testing Conclusions}
 }
