### 2021-01-24 #2826 @usethedata
@jGaboardi I've had a couple more declines on reviewers.  As it turns out, however, I'm stepping away from being a JOSS Topic editor and returning to just being a reviewer.  Ping me here when you've addressed, as you feel appropriate, the items @martibosch has listed, and I'll do the second review.
