### 2021-01-24 aec3089
update Network docstring
