+++ docs/index.rst
@@ -94,7 +94,7 @@ If you use PySAL-spaghetti in a scientific publication, we would appreciate usin
 Citing Work
 -----------
 
-  * **Lovelace, R**. `Open source tools for geographic analysis in transport planning`. J Geogr Syst (2021). https://doi.org/10.1007/s10109-020-00342-2.
+* **Lovelace, R**. `Open source tools for geographic analysis in transport planning`. J Geogr Syst (2021). https://doi.org/10.1007/s10109-020-00342-2.
 
 
 Funding
