### 2021-01-25 3c66176
add Moran method to Network
