+++ spaghetti/network.py
@@ -2,6 +2,7 @@ from collections import defaultdict, OrderedDict
 from itertools import islice
 import copy, os, pickle, warnings
 
+import esda
 import numpy
 
 from .analysis import GlobalAutoK
@@ -853,7 +854,6 @@ class Network:
 
         >>> import spaghetti
         >>> from libpysal import examples
-        >>> import esda
         >>> import numpy
         >>> ntw = spaghetti.Network(examples.get_path("streets.shp"))
 
@@ -874,37 +874,13 @@ class Network:
         Create a contiguity-based ``W`` object.
 
         >>> w = ntw.contiguityweights(graph=False)
-
-        Using the ``W`` object, access to `esda <https://pysal.org/esda/>`_
-        functionality is provided. First, a vector of attributes is
-        created for all edges with observations.
-
-        >>> w = ntw.contiguityweights(graph=False)
-        >>> arcs = w.neighbors.keys()
-        >>> y = numpy.zeros(len(arcs))
-        >>> for i, e in enumerate(arcs):
-        ...     if e in counts.keys():
-        ...         y[i] = counts[e]
-
-        Fetch the number of observations associated with arc ``3``,
-        where ``3`` is the basic 0-indexed ID of ``w.neighbors.keys()``
-        created through ``enumerate(arcs)``.
-
-        >>> y[3]
-        3.0
-
-        Next, a standard call to
-        `esda.Moran <https://pysal.org/esda/generated/esda.Moran.html#esda.Moran>`_
-        is made and the result placed into ``res``.
-
-        >>> res = esda.moran.Moran(y, w, permutations=99)
-        >>> type(res)
-        <class 'esda.moran.Moran'>
+        >>> w.n, w.n_components
+        (303, 1)
 
         Notes
         -----
 
-        See :cite:`pysal2007` and :cite:`esda:_2019` for more details.
+        See :cite:`pysal2007` for more details.
 
         """
 
@@ -1298,20 +1274,23 @@ class Network:
         pointpattern.dist_snapped = dist_snapped
         pointpattern.obs_to_vertex = list(obs_to_vertex)
 
-    def count_per_link(self, obs_on, graph=True):
+    def count_per_link(self, obs_on, graph=False):
         """Compute the counts per arc or edge (link).
 
         Parameters
         ----------
         obs_on : dict
             Dictionary of observations on the network.
-            Either in the form {(<LINK>):{<POINT_ID>:(<COORDS>)}} or
-            {<LINK>:[(<COORD>),(<COORD>)]}.
+            Either in the form ``{(<LINK>):{<POINT_ID>:(<COORDS>)}}``
+            or ``{<LINK>:[(<COORD>),(<COORD>)]}``.
+        graph : bool
+            Count observations on graph edges (``True``) or
+            network arcs (``False``). Default is ``False``.
 
         Returns
         -------
         counts : dict
-            Counts per network link in the form {(<LINK>):<COUNT>}.
+            Counts per network link in the form ``{(<LINK>):<COUNT>}``.
 
         Examples
         --------
@@ -2470,15 +2449,15 @@ class Network:
             A ``spaghetti`` point pattern object.
         nsteps : int
             The number of steps at which the count of the nearest
-            neighbors is computed. Default is 10.
+            neighbors is computed. Default is ``10``.
         permutations : int
-            The number of permutations to perform. Default is 99.
+            The number of permutations to perform. Default is ``99``.
         threshold : float
             The level at which significance is computed.
-            (0.5 would be 97.5% and 2.5%). Default is 0.5.
+            (0.5 would be 97.5% and 2.5%). Default is ``0.5``.
         distribution : str
             The distribution from which random points are sampled.
-            Currently, the only supported distribution is uniform.
+            Currently, the only supported distribution is ``'uniform'``.
         upperbound : float
[…]
