+++ spaghetti/tests/network_unittest_classes.py
@@ -699,6 +699,7 @@ class TestNetworkPointPattern(unittest.TestCase):
 # -------------------------------------------------------------------------------
 class TestNetworkAnalysis(unittest.TestCase):
     def setUp(self):
+        # synthetic test data
         bounds, h, v = (0, 0, 3, 3), 2, 2
         lattice = self.spaghetti.regular_lattice(bounds, h, nv=v, exterior=True)
         self.ntw = self.spaghetti.Network(in_data=lattice)
@@ -721,6 +722,10 @@ class TestNetworkAnalysis(unittest.TestCase):
         npts = self.ntw.pointpatterns[self.mids].npoints
         self.test_permutations = 99
         self.test_steps = 10
+
+        # empirical test_data
+        self.ntw_shp = self.spaghetti.Network(in_data=STREETS)
+        self.ntw_shp.snapobservations(CRIMES, crimes, attribute=True)
 
     def tearDown(self):
         pass
@@ -758,6 +763,18 @@ class TestNetworkAnalysis(unittest.TestCase):
                 nsteps=self.test_steps,
                 distribution="mrofinu",
             )
+
+    def test_moran_network(self):
+        known_moran_I, known_y = 0.005193, [0.0, 1, 0.0, 3, 0.0]
+        observed_moran, observed_y = self.ntw_shp.Moran(crimes)
+        self.assertAlmostEqual(observed_moran.I, known_moran_I, places=6)
+        self.assertEqual(observed_y[:5], known_y)
+
+    def test_moran_graph(self):
+        known_moran_I, known_y = 0.004778, [1, 0.0, 0.0, 3, 1]
+        observed_moran, observed_y = self.ntw_shp.Moran(crimes, graph=True)
+        self.assertAlmostEqual(observed_moran.I, known_moran_I, places=6)
+        self.assertEqual(observed_y[:5], known_y)
 
 
 # -------------------------------------------------------------------------------
