+++ spaghetti/tests/network_unittest_classes.py
@@ -1,4 +1,5 @@
 from libpysal import cg, examples, io
+from libpysal.common import RTOL, ATOL
 import numpy
 import unittest
 import copy
@@ -765,17 +766,19 @@ class TestNetworkAnalysis(unittest.TestCase):
             )
 
     def test_moran_network(self):
-        known_moran_I, known_y = 0.005193, [0.0, 1, 0.0, 3, 0.0]
-        numpy.random.seed(0)
+        known_moran_I, known_y = 0.005192687496078421, [0.0, 1, 0.0, 3, 0.0]
         observed_moran, observed_y = self.ntw_shp.Moran(crimes)
-        self.assertAlmostEqual(observed_moran.I, known_moran_I, places=6)
+        numpy.testing.assert_allclose(
+            observed_moran.I, known_moran_I, rtol=RTOL, atol=ATOL
+        )
         self.assertEqual(observed_y[:5], known_y)
 
     def test_moran_graph(self):
-        known_moran_I, known_y = 0.004778, [1, 0.0, 0.0, 3, 1]
-        numpy.random.seed(0)
+        known_moran_I, known_y = 0.004777863137379377, [1, 0.0, 0.0, 3, 1]
         observed_moran, observed_y = self.ntw_shp.Moran(crimes, graph=True)
-        self.assertAlmostEqual(observed_moran.I, known_moran_I, places=6)
+        numpy.testing.assert_allclose(
+            observed_moran.I, known_moran_I, rtol=RTOL, atol=ATOL
+        )
         self.assertEqual(observed_y[:5], known_y)
 
 
