### 2021-01-26 2c31f20
try test zero permutations
