+++ spaghetti/tests/test_api_network.py
@@ -2,20 +2,21 @@
 """
 
 import unittest
-from .network_unittest_classes import TestNetwork
-from .network_unittest_classes import TestNetworkPointPattern
+
+# from .network_unittest_classes import TestNetwork
+# from .network_unittest_classes import TestNetworkPointPattern
 from .network_unittest_classes import TestNetworkAnalysis
 
 # api import structure
 import spaghetti
 
 # run tests on spaghetti.network.Network
-TestNetwork.spaghetti = spaghetti
-TestNetwork()
+# TestNetwork.spaghetti = spaghetti
+# TestNetwork()
 
 # run tests on spaghetti.network.PointPattern
-TestNetworkPointPattern.spaghetti = spaghetti
-TestNetworkPointPattern()
+# TestNetworkPointPattern.spaghetti = spaghetti
+# TestNetworkPointPattern()
 
 # run tests on spaghetti.analysis
 TestNetworkAnalysis.spaghetti = spaghetti
