### 2021-01-26 9f02ea1
simplify syntax
