+++ .pre-commit-config.yaml
@@ -1,5 +1,5 @@
 repos: 
--   repo: https://github.com/ambv/black
+-   repo: https://github.com/psf/black
     rev: stable
     hooks: 
     - id: black
