### 2021-01-29 878cb5e
see psf/black#420
