+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos: 
 -   repo: https://github.com/psf/black
-    rev: stable
+    rev: 20.8b1
     hooks: 
     - id: black
       language_version: python3.9
