### 2021-01-29 8400267
Update sphinxcontrib-bibtex requirement from <2.0.0 to <3.0.0
Updates the requirements on [sphinxcontrib-bibtex](https://github.com/mcmtroffaes/sphinxcontrib-bibtex) to permit the latest version.
- [Release notes](https://github.com/mcmtroffaes/sphinxcontrib-bibtex/releases)
- [Changelog](https://github.com/mcmtroffaes/sphinxcontrib-bibtex/blob/develop/CHANGELOG.rst)
- [Commits](https://github.com/mcmtroffaes/sphinxcontrib-bibtex/compare/0.2...2.1.4)

Signed-off-by: dependabot[bot] <support@github.com>
