+++ requirements_docs.txt
@@ -1,5 +1,5 @@
 nbsphinx
 numpydoc
 sphinx>=1.4.3
-sphinxcontrib-bibtex<2.0.0
+sphinxcontrib-bibtex<3.0.0
 sphinx_bootstrap_theme
