### 2021-01-29 pysal/spaghetti#581 @jGaboardi
Partially addressed in #595.
