+++ .github/workflows/build_docs.yml
@@ -20,7 +20,7 @@
          shell: bash -l {0}
      steps:
        - uses: actions/checkout@v2
-       - uses: actions/cache@v2
+       - uses: actions/cache@v2.1.4
          env:
            CACHE_NUMBER: 0
          with:
