### 2021-03-16 #2826 @usethedata
Got it.  Personal issues can have an outsized effect these days.  Very much understood and I hope things have returned to a good state for you.
