### 2021-03-22 0552c9e
update Moran cite in docstring
