+++ .github/workflows/build_docs.yml
@@ -5,6 +5,12 @@
      # Sequence of patterns matched against refs/tags
      tags:
        - 'v*' # Push events to matching v*, i.e. v1.0, v20.15.10
+   workflow_dispatch:
+     inputs:
+       version:
+         description: Manual Doc Build Reason
+         default: test
+         required: false
  jobs:
    docs:
      name: Build & Push Docs
