+++ _sources/generated/spaghetti.Network.savenetwork.rst.txt
@@ -1,0 +1,6 @@
+﻿spaghetti.Network.savenetwork
+=============================
+
+.. currentmodule:: spaghetti
+
+.. automethod:: Network.savenetwork
