+++ _sources/generated/spaghetti.Network.loadnetwork.rst.txt
@@ -1,0 +1,6 @@
+﻿spaghetti.Network.loadnetwork
+=============================
+
+.. currentmodule:: spaghetti
+
+.. automethod:: Network.loadnetwork
