### 2021-04-02 pysal/spaghetti#574 @jGaboardi
We appreciate the suggestion that substantially improve the readability of our manuscript, and have made appropriate additions to the summary:

> Through `spaghetti`, first, network objects can be created and analysed from collections of line data by various means including reading in a shapefile or passing in a `geopandas.GeoDataFrame` at which time the line data are assigned network topology. Second, `spaghetti` provides computational tools to support statistical analysis of so-called network-based events along many different types of previously loaded networks. Network based-events or near-network observations are events that happen along spatial networks in our daily lives, i.e. locations of trees along footpaths, biking accidents along roads or locations of coffee shops along streets. As with `spaghetti.Network` objects, `spaghetti.PointPattern` objects can be [created from](https://pysal.org/spaghetti/generated/spaghetti.PointPattern.html#spaghetti.PointPattern) shapefiles, `geopandas.GeoDataFrame` objects or single `libpysal.cg.Point `objects. Near-network observations can then be snapped to nearest network segments enabling the calculation of observation distance matrices. Third, these observation distance matrices can be used both within `spaghetti` to perform clustering analysis or serve as input for other network-centric problems (e.g. optimal routing), or within the wider PySAL ecosystem to perform exploratory spatial analysis with [`esda`](https://pysal.org/esda/). Finally, `spaghetti`’s network elements (vertices and arcs) can also be extracted as `geopandas.GeoDataFrame` objects for visualization and integrated into further spatial statistical analysis within PySAL (e.g. `esda`).

We have also included a statement of need connected to the Related Works (see also Issue #575):

> While all the libraries above are important for network-based research, `spaghetti` was created and has evolved in line with the Python Spatial Analysis Library ecosystem for the specific purpose of utilizing the functionality of spatial weights in [`libpysal`](https://pysal.org/libpysal/) for generating network segment contiguity objects. The PySAL project was started in the mid-2000s when installation was difficult to maintain. Due to the non-triviality of relying on dependencies to secondary packages, a conscious decision was made to limit dependencies and build native PySAL data structures in cases where at all possible. Therefore, the original `pysal.network` submodule was developed to address the need for integrating support for network data structures with PySAL weights data structures, with the target audience being spatial data scientists and anyone interested in investigating network-centric phenomena within PySAL. Owing to the co-development of network functionality found within `spaghetti` and the evolution of the wider PySAL ecosystem, today, the package provides specialized network functionality that easily integrates with the rest of PySAL. This allows users of `spaghetti`’s network functionality to access spatial analysis functionality that complements network analysis, such as spatial statistical tools with `esda` and integration with core components of `libpysal`: `libpysal.weights` (mentioned above), `libpysal.cg` (computational geometry and data structures), `libpysal.io` (input-output), and `libpysal.examples` (built-in example data).
