### 2021-04-02 pysal/spaghetti#575 @jGaboardi
Thank you for the good suggestions to improve the manuscript!  We have incorporated information comprising a statement of need in the Related Works & Statement of Need section as suggested. As is also addressed in Issue #580, `spaghetti` is not a new Python package, but evolved from core PySAL functionality to a specialized submodule in the PySAL federation (originally `pysal.network`), which was in development prior to packages such OSMnx and concurrently with NetworkX. Importantly, package integration in the early spatial/scientific Python stack was problematic enough for the PySAL team to decide on “no external dependencies'' policy, which was a main driver in submodules like `pysal.network`. Owing to the co-development of network functionality found within `spaghetti` and the wider PySAL ecosystem, the package provides specialized network functionality that integrates with the rest of the PySAL ecosystem. This allows users of `spaghetti`’s network functionality to access spatial analysis functionality that complements network analysis, such as geovisualization with `splot` or other spatial statistical tools with `esda`.
