### 2021-04-02 pysal/spaghetti#576 @jGaboardi
Thank you for bringing this point. We agree with this suggestion. We have removed the image in #588.
