### 2021-04-02 pysal/spaghetti#577 @jGaboardi
We have updated the Quickstart tutorial (see #578) and will use it as an example in the future for improving other currently existing tutorials. This issue is tagged in #604 for our future work/wishlist.
