### 2021-04-02 pysal/spaghetti#578 @jGaboardi
Thank you for this valuable suggestion. We have addressed this issue by making the suggested changes in the Quickstart notebook. See also #603.
