### 2021-04-02 pysal/spaghetti#580 @jGaboardi
We thank the reviewer for making these points. We have addressed them in paragraph 2 in the second section:

> While all the libraries above are important for network-based research, `spaghetti` was created and has evolved in line with the Python Spatial Analysis Library ecosystem for the specific purpose of utilizing the functionality of spatial weights generated in [`libpysal`](https://pysal.org/libpysal/) for generating network segment contiguity objects. The PySAL project was started in the mid-2000s when installation was difficult to maintain. Due to the non-triviality of relying on dependencies to secondary packages, a conscious decision was made to limit dependencies and build native PySAL data structures in cases where at all possible. Therefore, the original `pysal.network` submodule was developed to address the need for integrating support for network data structures with PySAL weights data structures, with the target audience being spatial data scientists and anyone interested in investigating network-centric phenomena within PySAL. Owing to the co-development of network functionality found within `spaghetti` and the evolution of the wider PySAL ecosystem, today, the package provides specialized network functionality that easily integrates with the rest of PySAL. This allows users of `spaghetti`’s network functionality to access spatial analysis functionality that complements network analysis, such as spatial statistical tools with [`esda`](https://pysal.org/esda/) and integration with core components of `libpysal`: `libpysal.weights` (mentioned above), `libpysal.cg` (computational geometry and data structures), `libpysal.io` (input-output), and `libpysal.examples` (built-in example data).

Additionally, we provide more insight in #574 and #575.
