### 2021-04-02 pysal/spaghetti#581 @jGaboardi
Thank you for this remark, we have started to address this by adding a `Network_Moran` method (see the Network-constrained spatial autocorrelation notebook) into the code base as this is a much appreciated idea and we will be implementing that for this review. See Issue #586. Additionally, it is a good point that there is a fair amount of adhoc code, especially for plotting, in the tutorials. To stay in line with the federated ecosystem structure in the PySAL ecosystem, we plan on making general plotting functionality part of `pysal/splot` in the future. See issue https://github.com/pysal/splot/issues/108. Additionally, the code for optimization problems will become part of the `pysal/spopt` package following its initial release.
