### 2021-04-02 pysal/spaghetti#582 @jGaboardi
Thank you for pointing this out, we updated the documentation to be more clear (see #592)

> str, iterable (list, tuple, numpy.ndarray), libpysal.cg.Chain, geopandas.GeoDataFrame

We also added a wishlist/roadmap item to the issues (#593) to address reading these different formats in future. However, due to spaghetti being tightly integrated into the PySAL ecosystem, we believe that the majority of use cases are covered by the standard pysal input file formats such as shapefiles and GeoDataFrames.
