### 2021-04-02 pysal/spaghetti#583 @jGaboardi
Thank you for your suggestion, we extended the Current functionality section of the manuscript to more accurately and thoroughly describe what is present in `spaghetti`. The basic underlying process for creating a network instance is present (e.g., “2. generate the network representation”), as well as a textual description of secondary functions For example:
> * allocating observation point patterns to the network

and

> * utilizing observation counts on network segments and network spatial weights within the Moran’s *I* attribute to analyze global spatial autocorrelation”

Moreover, the code snippets for the two plots demonstrate and visualize the functionality associated with `spaghetti`, its parent package (`libpysal`), a required sibling package (`esda`), and optional dependencies. While the plotting is performed with `matplotlib` through `geopandas`, we believe visualization is a powerful tool in science for demonstration and should be retained here. As mentioned above, we decided to include another figure to demonstrate functionality since we are removing Figure 1 (the PySAL Logo) as per #576. This new figure can be found in #576/#606 with the associated caption as follows:

> Demonstrating the creation of a network and point pattern from shapefiles, followed by spatial autocorrelation analysis. A shapefile of school locations (blue) is read in and the points are snapped to the nearest network segments (green). A Moran's *I* statistic of -0.026 indicates near complete spatial randomness, though slightly dispersed.
