### 2021-04-02 04e56c8
update JOSS paper cite and figure size
