### 2021-04-02 bbfe5f2
revert figure sizes
