### 2021-04-02 #2826 @jGaboardi
@martibosch @usethedata @arfon We have wrapped up our responses to the issues raised by @martibosch, and implemented changes in the `spaghetti` package itself and in the manuscript text. I was unsure about leaving the issues opened by @martibosch as they are or closing them, so I left them open.

On another topic, I was unable to figure out how to reduce figure sizes elegantly in MarkDown, but I think both figures in the manuscript need to be smaller (pending your review & decision).

@slumnitz @sjsrey
