### 2021-04-03 7d331a6
clean up JOSS bib refs and cites
