+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{pysal2007,
-  author={Rey, Sergio J. and Anselin, Luc},
+  author={Sergio J. Rey and Luc Anselin},
   title={{PySAL: A Python Library of Spatial Analytical Methods}},
   journal={The Review of Regional Studies},
   year=2007,
@@ -7,245 +7,198 @@
   number={1},
   pages={5-27},
   keywords={Open Source; Software; Spatial},
-  url = {https://rrs.scholasticahq.com/article/8285.pdf}
+  url={https://rrs.scholasticahq.com/article/8285.pdf}
 }
 
 @article{pysal2015,
-  Title                    = {Open {Geospatial} {Analytics} with {PySAL}},
-  Author                   = {Rey, Sergio J. and Anselin, Luc and Li, Xun and Pahle, Robert and Laura, Jason and Li, Wenwen and Koschinsky, Julia},
-  Journal                  = {{ISPRS International Journal of Geo-Information}},
-  Year                     = {2015},
-  Number                   = {2},
-  Pages                    = {815--836},
-  Volume                   = {4},
-  Keywords                 = {open science},
-  doi = {10.3390/ijgi4020815}
+  title={Open {Geospatial} {Analytics} with {PySAL}},
+  author={Sergio J. Rey and Luc Anselin and Xun Li and Robert Pahle and Jason Laura and Wenwen Li and Julia Koschinsky},
+  journal={{ISPRS International Journal of Geo-Information}},
+  year={2015},
+  number={2},
+  pages={815--836},
+  volume={4},
+  keywords={open science},
+  doi={10.3390/ijgi4020815}
 }
 
 @article{gaboardi2020a,
-    author = {Gaboardi, James D. and Folch, David C. and Horner, Mark W.},
-    title = {{Connecting Points to Spatial Networks: Effects on Discrete Optimization Models}},
-    journal = {Geographical Analysis},
-    year = {2020},
-    volume = {52},
-    issue = {2},
-    pages = {299--322},
-    doi = {10.1111/gean.12211},
+  author={James D. Gaboardi and David C. Folch and Mark W. Horner},
+  title={{Connecting Points to Spatial Networks: Effects on Discrete Optimization Models}},
+  journal={Geographical Analysis},
+  year={2020},
+  volume={52},
+  issue={2},
+  pages={299--322},
+  doi={10.1111/gean.12211},
 }
 
 @misc{gaboardi2018,
-    author = {Gaboardi, James D. and Laura, Jay and Rey, Sergio and 
-                 Wolf, Levi John and Folch, David C. and Kang, Wei and 
-                 Stephens, Philip and Schmidt, Charles},
-    month     = {oct},
-    year      = {2018},
-    title     = {pysal/spaghetti},
-    url       = {https://github.com/pysal/spaghetti},
-    doi       = {10.5281/zenodo.1343650},
-    keywords  = {graph-theory,network-analysis,python,spatial-networks,topology}
+  author={James D. Gaboardi and Jay Laura and Sergio Rey and Levi John Wolf and David C. Folch and Wei Kang and Philip Stephens and Charles Schmidt},
+  month={oct},
+  year={2018},
+  title={pysal/spaghetti},
+  url= {https://github.com/pysal/spaghetti},
+  doi={10.5281/zenodo.1343650},
+  keywords={graph-theory,network-analysis,python,spatial-networks,topology}
 }
 
 @article{okabe2006a,
-    author = {Okabe, Atsuyuki and Okunuki, Keiichi and Shiode, Shino},
-    doi = {10.1111/j.0016-7363.2005.00674.x},
-    isbn = {1538-4632},
[…]
