### 2021-04-07 #2826 @arfon
> On another topic, I was unable to figure out how to reduce figure sizes elegantly in MarkDown, but I think both figures in the manuscript need to be smaller (pending your review & decision).

@jGaboardi – Figure sizes can be customized by adding an optional second parameter:

```
![Caption for example figure.](figure.png){ width=20% }
```
