### 2021-04-07 d82dbec
reduce JOSS figures by 50%
