### 2021-04-07 4638f5d
clean up JOSS bib refs and cites [2]
