+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{pysal2007,
-  author={Sergio J. Rey and Luc Anselin},
+  author={Sergio Rey and Luc Anselin},
   title={{PySAL: A Python Library of Spatial Analytical Methods}},
   journal={The Review of Regional Studies},
   year=2007,
@@ -12,7 +12,7 @@
 
 @article{pysal2015,
   title={Open {Geospatial} {Analytics} with {PySAL}},
-  author={Sergio J. Rey and Luc Anselin and Xun Li and Robert Pahle and Jason Laura and Wenwen Li and Julia Koschinsky},
+  author={Sergio Rey and Luc Anselin and Xun Li and Robert Pahle and Jason Laura and Wenwen Li and Julia Koschinsky},
   journal={{ISPRS International Journal of Geo-Information}},
   year={2015},
   number={2},
@@ -130,7 +130,7 @@
 }
 
 @misc{libpysal2020,
-  author={Sergio Rey and Philip Stephens and Levi John Wolf and Charles Schmidt and jlaura and Taylor Oshan and Dani Arribas-Bel and James Gaboardi and David C. Folch and mhwang4 and Wei Kang and Nicholas Malizia and Pedro Amaral and Luc Anselin and Eli Knaap and Hu Shao and Marynia and Andrew Winslow and Conceptron and Jotham Apaloo and Martin Fleischmann and Andy Eschbacher and Stefanie Lumnitz and Siddharth S. and Forest Gregg and Elliott Sales de Andrade and Bas Couwenberg and Josh Kalderimis and Hannes and Andy Reagan},
+  author={Sergio Rey and Philip Stephens and Levi John Wolf and Charles Schmidt and jlaura and Taylor Oshan and Dani Arribas-Bel and James Gaboardi and David Folch and mhwang4 and Wei Kang and Nicholas Malizia and Pedro Amaral and Luc Anselin and Eli Knaap and Hu Shao and Marynia and Andrew Winslow and Conceptron and Jotham Apaloo and Martin Fleischmann and Andy Eschbacher and Stefanie Lumnitz and Siddharth S and Forest Gregg and Elliott Sales de Andrade and Bas Couwenberg and Josh Kalderimis and Hannes and Andy Reagan},
   title={pysal/libpysal: v4.2.2},
   month=feb,
   year=2020,
@@ -161,7 +161,7 @@
 }
 
 @inbook{osullivan_unwin2010.ch5,
-  author={D. O'Sullivan and D.J. Unwin,},
+  author={David O'Sullivan and David J. Unwin},
   publisher={John Wiley \& Sons, Ltd},
   isbn={9780470549094},
   title={Point Pattern Analysis},
@@ -186,7 +186,7 @@
 }
 
 @misc{esda:_2019,
-  author={S.J. Rey and L.J. Wolf and W. Kang and P. Stephens and J. Laura and C. Schmidt and D. Arribas-Bel and S. Lumnitz and J.C. Duque and D.C. Folch and L. Anselin and N. Malizia and J.D. Gaboardi and F. Fernandes and M. Seth and mhwang4 and mlyons-tcc},
+  author={Sergio Rey and Levi Wolf and Wei Kang and Phil Stephens and Jason Laura and Charles Schmidt and Dani Arribas-Bel and Stefanie Lumnitz and Juan Carlos Duque and David Folch and Luc Anselin and Nicholas Malizia and James Gaboardi and Filipe Fernandes and Mridul Seth and mhwang4 and mlyons-tcc},
   title={{pysal/esda}},
   month=jul,
   year={2019},
@@ -197,7 +197,7 @@
 
 @book{moran:_cliff81,
   address={London},
[…]
