### 2021-04-09 #2826 @jGaboardi
I'll go ahead and close the issues that @martibosch opened as we are "done" with them. If I am jumping the gun here, I'm happy to reopen.

[#574](https://github.com/pysal/spaghetti/issues/574), [#575](https://github.com/pysal/spaghetti/issues/575), [#576](https://github.com/pysal/spaghetti/issues/576), [#577](https://github.com/pysal/spaghetti/issues/577), [#578](https://github.com/pysal/spaghetti/issues/578), [#580](https://github.com/pysal/spaghetti/issues/580), [#581](https://github.com/pysal/spaghetti/issues/581), [#582](https://github.com/pysal/spaghetti/issues/582), [#583](https://github.com/pysal/spaghetti/issues/583)
