+++ requirements.txt
@@ -2,5 +2,5 @@ esda
 libpysal
 numpy>=1.3
 pandas>=1.0
-rtree
+rtree==0.9.3
 scipy>=1.0
