+++ ci/39-PYPI.yaml
@@ -8,7 +8,7 @@ dependencies:
   - esda
   - numpy
   - pandas>=1.0
-  - rtree
+  - rtree=0.9.3
   - scipy>=1.0
   - watermark
   # testing/formatting
