### 2021-04-21 dc80d1e
remove mamba & conda mixture
