+++ .github/workflows/unittests.yml
@@ -36,7 +36,6 @@
        - uses: conda-incubator/setup-miniconda@v2
          with:
             miniconda-version: 'latest'
-            mamba-version: '*'
             channels: conda-forge
             channel-priority: true
             auto-update-conda: false
@@ -44,8 +43,8 @@
             environment-file: ${{ matrix.environment-file }}
             activate-environment: test
             use-only-tar-bz2: true
-       - run: mamba info --all
-       - run: mamba list
+       - run: conda info --all
+       - run: conda list
        - run: conda config --show-sources
        - run: conda config --show
        - run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
