### 2021-04-22 cbc8cc0
try mamba + strict again
