### 2021-04-22 8961ae8
remove unsupported mamba commands
