+++ .github/workflows/unittests.yml
@@ -10,7 +10,7 @@
 
  jobs:
    unittests:
-     name: conda (${{ matrix.os }}, ${{ matrix.environment-file }})
+     name: mamba (${{ matrix.os }}, ${{ matrix.environment-file }})
      runs-on: ${{ matrix.os }}
      timeout-minutes: 30
      strategy:
@@ -47,8 +47,8 @@
             use-only-tar-bz2: true
        - run: mamba info --all
        - run: mamba list
-       - run: mamba config --show-sources
-       - run: mamba config --show
+       #- run: mamba config --show-sources
+       #- run: mamba config --show
        - run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
        - name: codecov (${{ matrix.os }}, ${{ matrix.environment-file }})
          uses: codecov/codecov-action@v1
