+++ .github/workflows/unittests.yml
@@ -8,45 +8,33 @@
    schedule:
      - cron: '59 23 * * *'
 
- jobs:
-   unittests:
-     name: conda (${{ matrix.os }}, ${{ matrix.environment-file }})
-     runs-on: ${{ matrix.os }}
-     timeout-minutes: 30
-     strategy:
-       matrix:
-         os: [macos-latest, ubuntu-latest, windows-latest]
-         environment-file: [ci/36-PYPI.yaml, ci/36-PYPI-PLUS.yaml, ci/36-GIT.yaml, ci/36-GIT-PLUS.yaml, ci/37-PYPI.yaml, ci/37-PYPI-PLUS.yaml, ci/37-GIT.yaml, ci/37-GIT-PLUS.yaml, ci/38-PYPI.yaml, ci/38-PYPI-PLUS.yaml, ci/38-GIT.yaml, ci/38-GIT-PLUS.yaml, ci/39-PYPI.yaml, ci/39-PYPI-PLUS.yaml, ci/39-GIT.yaml, ci/39-GIT-PLUS.yaml]
-         exclude:
-           - environment-file: ci/36-PYPI-PLUS.yaml
-             os: windows-latest
-           - environment-file: ci/36-GIT-PLUS.yaml
-             os: windows-latest
-     defaults:
-       run:
-         shell: bash -l {0}
-     steps:
-       - uses: actions/checkout@v2
-       - uses: actions/cache@v2
-         env:
-           CACHE_NUMBER: 0
-         with:
-           path: ~/conda_pkgs_dir
-           key: ${{ matrix.os }}-conda-${{ env.CACHE_NUMBER }}-${{ hashFiles(matrix.environment-file) }}
-       - uses: conda-incubator/setup-miniconda@v2
-         with:
-            miniconda-version: 'latest'
-            channels: conda-forge
-            channel-priority: strict
-            auto-update-conda: false
-            auto-activate-base: false
-            environment-file: ${{ matrix.environment-file }}
-            activate-environment: test
-            use-only-tar-bz2: true
-       - run: conda info --all
-       - run: conda list
-       - run: conda config --show-sources
-       - run: conda config --show
+jobs:
+  unittests:
+    name: conda (${{ matrix.os }}, ${{ matrix.environment-file }})
+    runs-on: ${{ matrix.os }}
+    timeout-minutes: 30
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+        #os: [macos-latest, ubuntu-latest, windows-latest]
+        environment-file: [ci/39-PYPI.yaml, ci/39-PYPI-PLUS.yaml, ci/39-GIT.yaml, ci/39-GIT-PLUS.yaml]
+        #environment-file: [ci/36-PYPI.yaml, ci/36-PYPI-PLUS.yaml, ci/36-GIT.yaml, ci/36-GIT-PLUS.yaml, ci/37-PYPI.yaml, ci/37-PYPI-PLUS.yaml, ci/37-GIT.yaml, ci/37-GIT-PLUS.yaml, ci/38-PYPI.yaml, ci/38-PYPI-PLUS.yaml, ci/38-GIT.yaml, ci/38-GIT-PLUS.yaml, ci/39-PYPI.yaml, ci/39-PYPI-PLUS.yaml, ci/39-GIT.yaml, ci/39-GIT-PLUS.yaml]
+        #exclude:
+        #  - environment-file: ci/36-PYPI-PLUS.yaml
+        #    os: windows-latest
+        #  - environment-file: ci/36-GIT-PLUS.yaml
+        #    os: windows-latest
+    defaults:
+      run:
+        shell: bash -l {0}
+    steps:
+      - uses: actions/checkout@v2
+      - uses: mamba-org/provision-with-micromamba@main
+        with:
[…]
