+++ .github/workflows/unittests.yml
@@ -31,8 +31,8 @@
        - uses: actions/checkout@v2
        - uses: mamba-org/provision-with-micromamba@main
          with:
-           environment-file: ${{ matrix.environment-file }}
-           environment-name: test
+           - environment-file: ${{ matrix.environment-file }}
+           - environment-name: test
         - run: mamba info --all
         - run: mamba list
         - run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
