+++ .github/workflows/unittests.yml
@@ -10,20 +10,20 @@
 
  jobs:
    unittests:
-     name: micromamba (${{ matrix.os }}, ${{ matrix.environment-file }})
-     runs-on: ${{ matrix.os }}
+     name: micromamba (${{ ubuntu-latest }}, ${{ ci/39-PYPI.yaml }})
+     runs-on: ubuntu-latest
      timeout-minutes: 30
-     strategy:
-       matrix:
-         os: [ubuntu-latest]
-         #os: [macos-latest, ubuntu-latest, windows-latest]
-         environment-file: [ci/39-PYPI.yaml, ci/39-PYPI-PLUS.yaml, ci/39-GIT.yaml, ci/39-GIT-PLUS.yaml]
-         #environment-file: [ci/36-PYPI.yaml, ci/36-PYPI-PLUS.yaml, ci/36-GIT.yaml, ci/36-GIT-PLUS.yaml, ci/37-PYPI.yaml, ci/37-PYPI-PLUS.yaml, ci/37-GIT.yaml, ci/37-GIT-PLUS.yaml, ci/38-PYPI.yaml, ci/38-PYPI-PLUS.yaml, ci/38-GIT.yaml, ci/38-GIT-PLUS.yaml, ci/39-PYPI.yaml, ci/39-PYPI-PLUS.yaml, ci/39-GIT.yaml, ci/39-GIT-PLUS.yaml]
-         #exclude:
-         #  - environment-file: ci/36-PYPI-PLUS.yaml
-         #    os: windows-latest
-         #  - environment-file: ci/36-GIT-PLUS.yaml
-         #    os: windows-latest
+     #strategy:
+     #  matrix:
+     #    os: [ubuntu-latest]
+     #    #os: [macos-latest, ubuntu-latest, windows-latest]
+     #    environment-file: [ci/39-PYPI.yaml, ci/39-PYPI-PLUS.yaml, ci/39-GIT.yaml, ci/39-GIT-PLUS.yaml]
+     #    #environment-file: [ci/36-PYPI.yaml, ci/36-PYPI-PLUS.yaml, ci/36-GIT.yaml, ci/36-GIT-PLUS.yaml, ci/37-PYPI.yaml, #ci/37-PYPI-PLUS.yaml, ci/37-GIT.yaml, ci/37-GIT-PLUS.yaml, ci/38-PYPI.yaml, ci/38-PYPI-PLUS.yaml, #ci/38-GIT.yaml, ci/38-GIT-PLUS.yaml, ci/39-PYPI.yaml, ci/39-PYPI-PLUS.yaml, ci/39-GIT.yaml, #ci/39-GIT-PLUS.yaml]
+     #    #exclude:
+     #    #  - environment-file: ci/36-PYPI-PLUS.yaml
+     #    #    os: windows-latest
+     #    #  - environment-file: ci/36-GIT-PLUS.yaml
+     #    #    os: windows-latest
      defaults:
        run:
          shell: bash -l {0}
@@ -31,12 +31,12 @@
        - uses: actions/checkout@v2
        - uses: mamba-org/provision-with-micromamba@main
          with:
-           - environment-file: ${{ matrix.environment-file }}
+           - environment-file: ci/39-PYPI.yaml
            - environment-name: test
        - run: mamba info --all
        - run: mamba list
        - run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
-       - name: codecov (${{ matrix.os }}, ${{ matrix.environment-file }})
+       - name: codecov (${{ ubuntu-latest }}, ${{ ci/39-PYPI.yaml }})
          uses: codecov/codecov-action@v1
          with:
            token: ${{ secrets.CODECOV_TOKEN }}
