+++ .github/workflows/unittests.yml
@@ -32,7 +32,6 @@
        - uses: mamba-org/provision-with-micromamba@main
          with:
            - environment-file: ci/39-PYPI.yaml
-           - environment-name: test
        - run: mamba info --all
        - run: mamba list
        - run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
