+++ .github/workflows/unittests.yml
@@ -18,10 +18,10 @@
          shell: bash -l {0}
      steps:
        - uses: actions/checkout@v2
-       
-       - uses: mamba-org/provision-with-micromamba@main
+       - name: somename
+         uses: mamba-org/provision-with-micromamba@main
          with:
-           - environment-file: ci/39-PYPI.yaml
+           environment-file: ci/39-PYPI.yaml
        - run: mamba info --all
        - run: mamba list
        - run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
