+++ .github/workflows/unittests.yml
@@ -10,6 +10,55 @@
 
  jobs:
    unittests:
+     name: micromamba (${{ matrix.os }}, ${{ matrix.environment-file }})
+     runs-on: ${{ matrix.os }}
+     timeout-minutes: 30
+     strategy:
+       matrix:
+         os: [macos-latest, ubuntu-latest, windows-latest]
+         environment-file: [ci/36-PYPI.yaml, ci/36-PYPI-PLUS.yaml, ci/36-GIT.yaml, ci/36-GIT-PLUS.yaml, ci/37-PYPI.yaml,  ci/37-PYPI-PLUS.yaml, ci/37-GIT.yaml, ci/37-GIT-PLUS.yaml, ci/38-PYPI.yaml, ci/38-PYPI-PLUS.yaml,  ci/38-GIT.yaml, ci/38-GIT-PLUS.yaml, ci/39-PYPI.yaml, ci/39-PYPI-PLUS.yaml, ci/39-GIT.yaml,  ci/39-GIT-PLUS.yaml]
+         exclude:
+           - environment-file: ci/36-PYPI-PLUS.yaml
+             os: windows-latest
+           - environment-file: ci/36-GIT-PLUS.yaml
+             os: windows-latest
+     defaults:
+       run:
+         shell: bash -l {0}
+     steps:
+       - uses: actions/checkout@v2
+       - name: install mamba
+         uses: mamba-org/provision-with-micromamba@main
+         with:
+           environment-file: ${{ matrix.environment-file }}
+       - name: run tests
+         run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
+       - name: codecov (${{ matrix.os }}, ${{ matrix.environment-file }})
+         uses: codecov/codecov-action@v1
+         with:
+           token: ${{ secrets.CODECOV_TOKEN }}
+           file: ./coverage.xml
+           name: spaghetti-codecov
+
+ 
+ 
+ 
+ 
+ 
+ 
+ 
+ 
+ 
+ 
+ 
+ 
+ 
+ 
+ 
+ 
+ 
+ jobs:
+   unittests:
      name: micromamba
      runs-on: ubuntu-latest
      timeout-minutes: 30
