+++ .github/workflows/unittests.yml
@@ -40,47 +40,32 @@
            file: ./coverage.xml
            name: spaghetti-codecov
 
+
  
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- jobs:
-   unittests:
-     name: micromamba
-     runs-on: ubuntu-latest
-     timeout-minutes: 30
-     defaults:
-       run:
-         shell: bash -l {0}
-     steps:
-       - uses: actions/checkout@v2
-       - name: install mambda
-         uses: mamba-org/provision-with-micromamba@main
-         with:
-           environment-file: ci/39-PYPI.yaml
-       #- run: mamba info --all
-       #- run: mamba list
-       - name: run tests
-         run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
-       - name: codecov
-         uses: codecov/codecov-action@v1
-         with:
-           token: ${{ secrets.CODECOV_TOKEN }}
-           file: ./coverage.xml
-           name: spaghetti-codecov
+# jobs:
+#   unittests:
+#     name: micromamba
+#     runs-on: ubuntu-latest
+#     timeout-minutes: 30
+#     defaults:
+#       run:
+#         shell: bash -l {0}
+#     steps:
+#       - uses: actions/checkout@v2
+#       - name: install mambda
+#         uses: mamba-org/provision-with-micromamba@main
+#         with:
+#           environment-file: ci/39-PYPI.yaml
+#       #- run: mamba info --all
+#       #- run: mamba list
+#       - name: run tests
+#         run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
+#       - name: codecov
+#         uses: codecov/codecov-action@v1
+#         with:
+#           token: ${{ secrets.CODECOV_TOKEN }}
+#           file: ./coverage.xml
+#           name: spaghetti-codecov
 
 
 #jobs:
