+++ .github/workflows/unittests.yml
@@ -32,8 +32,14 @@
          uses: mamba-org/provision-with-micromamba@main
          with:
            environment-file: ${{ matrix.environment-file }}
-       - name: run tests
+       - name: run tests in bash
+         shell: bash -l {0}
          run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
+         if matrix.os != 'Windows'
+       - name: run tests in powershell
+         shell: powershell
+         run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
+         if matrix.os == 'Windows'
        - name: codecov (${{ matrix.os }}, ${{ matrix.environment-file }})
          uses: codecov/codecov-action@v1
          with:
