### 2021-04-24 0042c6b
os value for if logic
