+++ .github/workflows/unittests.yml
@@ -35,11 +35,11 @@
        - name: run tests in bash
          shell: bash -l {0}
          run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
-         if: matrix.os != 'Windows'
+         if: matrix.os != 'windows-latest'
        - name: run tests in powershell
          shell: powershell
          run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
-         if: matrix.os == 'Windows'
+         if: matrix.os == 'windows-latest'
        - name: codecov (${{ matrix.os }}, ${{ matrix.environment-file }})
          uses: codecov/codecov-action@v1
          with:
