### 2021-04-24 9dac78c
try PYPI vs. GIT logic
