### 2021-04-24 8533f78
try PYPI vs. GIT logic [2]
