### 2021-04-24 0cebd8c
try PYPI vs. GIT logic [3]
