### 2021-04-24 7c0cb61
try PYPI vs. GIT logic [4]
