+++ .github/workflows/unittests.yml
@@ -32,7 +32,7 @@
          uses: mamba-org/provision-with-micromamba@main
          with:
            environment-file: ${{ matrix.environment-file }}
-           if: matrix.environment-file == 'ci/*PYPI*.yaml'
+         if: matrix.environment-file == 'ci/*PYPI*.yaml'
          - name: run tests in bash
            shell: bash -l {0}
            run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
