### 2021-04-24 4c3da8e
try PYPI vs. GIT logic [5]
