+++ .github/workflows/unittests.yml
@@ -32,15 +32,15 @@
          uses: mamba-org/provision-with-micromamba@main
          with:
            environment-file: ${{ matrix.environment-file }}
-         if: matrix.environment-file == 'ci/*PYPI*.yaml'
-         - name: run tests in bash
-           shell: bash -l {0}
-           run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
-           if: matrix.os != 'windows-latest'
-         - name: run tests in powershell
-           shell: powershell
-           run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
-           if: matrix.os == 'windows-latest'
+         if: contains(matrix.environment-file, 'PYPI')
+       - name: run tests in bash
+         shell: bash -l {0}
+         run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
+         if: matrix.os != 'windows-latest' && contains(matrix.environment-file, 'PYPI')
+       - name: run tests in powershell
+         shell: powershell
+         run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
+         if: matrix.os == 'windows-latest' && contains(matrix.environment-file, 'PYPI')
        - name: codecov (${{ matrix.os }}, ${{ matrix.environment-file }})
          uses: codecov/codecov-action@v1
          with:
