### 2021-04-24 1b96bf7
try PYPI vs. GIT logic [6]
