### 2021-04-24 1827dd2
try PYPI vs. GIT logic [7]
