### 2021-04-24 0593b49
try PYPI vs. GIT logic [8]
