### 2021-04-24 dd573f3
try PYPI vs. GIT logic [9]
