### 2021-04-25 c7c3620
try PYPI vs. GIT logic [12]
