### 2021-04-25 c74a6bc
try PYPI vs. GIT logic [13]
