### 2021-04-25 b08222e
try PYPI vs. GIT logic [14]
