### 2021-04-25 dc7cf81
try PYPI vs. GIT logic [15]
