### 2021-04-25 eac021d
try PYPI vs. GIT logic [17]
