### 2021-04-25 9c87f17
try PYPI vs. GIT logic [19]
