### 2021-04-25 96a9afd
try @martinfleis's libpysal idea
