### 2021-04-25 c16a53d
remove mamba vs. micromamba logic
