### 2021-04-25 6235c31
streamline/correct logic
