+++ .github/workflows/unittests.yml
@@ -16,8 +16,8 @@
      strategy:
        matrix:
          os: [macos-latest, ubuntu-latest, windows-latest]
-         environment-file: [ci/39-PYPI.yaml, ci/39-PYPI-PLUS.yaml, ci/39-GIT.yaml, ci/39-GIT-PLUS.yaml]
-         #environment-file: [ci/36-PYPI.yaml, ci/36-PYPI-PLUS.yaml, ci/36-GIT.yaml, ci/36-GIT-PLUS.yaml, ci/37-PYPI.yaml, ci/37-PYPI-PLUS.yaml, ci/37-GIT.yaml, ci/37-GIT-PLUS.yaml, ci/38-PYPI.yaml, ci/38-PYPI-PLUS.yaml, ci/38-GIT.yaml, ci/38-GIT-PLUS.yaml, ci/39-PYPI.yaml, ci/39-PYPI-PLUS.yaml, ci/39-GIT.yaml, ci/39-GIT-PLUS.yaml]
+         #environment-file: [ci/39-PYPI.yaml, ci/39-PYPI-PLUS.yaml, ci/39-GIT.yaml, ci/39-GIT-PLUS.yaml]
+         environment-file: [ci/36-PYPI.yaml, ci/36-PYPI-PLUS.yaml, ci/36-GIT.yaml, ci/36-GIT-PLUS.yaml, ci/37-PYPI.yaml, ci/37-PYPI-PLUS.yaml, ci/37-GIT.yaml, ci/37-GIT-PLUS.yaml, ci/38-PYPI.yaml, ci/38-PYPI-PLUS.yaml, ci/38-GIT.yaml, ci/38-GIT-PLUS.yaml, ci/39-PYPI.yaml, ci/39-PYPI-PLUS.yaml, ci/39-GIT.yaml, ci/39-GIT-PLUS.yaml]
          exclude:
            - environment-file: ci/36-PYPI-PLUS.yaml
              os: windows-latest
@@ -32,7 +32,6 @@
          with:
            environment-file: ${{ matrix.environment-file }}
            micromamba-version: 'latest'
-         #if: contains(matrix.environment-file, 'PYPI')
       
        - name: install bleeding edge libpysal - bash
          shell: bash -l {0}
@@ -48,34 +47,15 @@
            pip install git+https://github.com/pysal/libpysal.git@master
          if: matrix.os == 'windows-latest' && contains(matrix.environment-file, 'GIT')
        
-       #- name: setup mamba
-       #  uses: conda-incubator/setup-miniconda@v2
-       #  with:
-       #    miniconda-version: 'latest'
-       #    mamba-version: '*'
-       #    use-mamba: true
-       #    channels: conda-forge
-       #    channel-priority: strict
-       #    auto-update-conda: false
-       #    auto-activate-base: false
-       #    environment-file: ${{ matrix.environment-file }}
-       #    activate-environment: test
-       #  if: contains(matrix.environment-file, 'GIT')
-       
-       - name: run tests - bash PYPI
+       - name: run tests - bash
          shell: bash -l {0}
          run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
-         if: matrix.os != 'windows-latest' && contains(matrix.environment-file, 'PYPI')
-       
-       - name: run tests - bash GIT
-         shell: bash -l {0}
-         run: pytest -v spaghetti --cov=spaghetti --doctest-modules --cov-config=.coveragerc --cov-report=xml
-         if: contains(matrix.environment-file, 'GIT')
+         if: matrix.os != 'windows-latest'
        
        - name: run tests - powershell
          shell: powershell
[…]
