### 2021-04-25 bcb375a
enusre 39 is good before getting fancy
