+++ ci/38-GIT-PLUS.yaml
@@ -17,7 +17,3 @@ dependencies:
   - pytest-cov
   # optional
   - geopandas>=0.7.0
-  # required
-  - pip
-  - pip:
-    - git+https://github.com/pysal/libpysal.git@master
