+++ ci/38-GIT.yaml
@@ -15,7 +15,3 @@ dependencies:
   - codecov
   - pytest
   - pytest-cov
-  # required
-  - pip
-  - pip:
-    - git+https://github.com/pysal/libpysal.git@master
