### 2021-04-25 25aaa95
rename PYPI/GIT schema
