+++ .github/workflows/unittests.yml
@@ -16,12 +16,13 @@
      strategy:
        matrix:
          os: [macos-latest, ubuntu-latest, windows-latest]
-         environment-file: [ci/36-PYPI.yaml, ci/36-PYPI-PLUS.yaml, ci/36-GIT.yaml, ci/36-GIT-PLUS.yaml, ci/37-PYPI.yaml, ci/37-PYPI-PLUS.yaml, ci/37-GIT.yaml, ci/37-GIT-PLUS.yaml, ci/38-PYPI.yaml, ci/38-PYPI-PLUS.yaml, ci/38-GIT.yaml, ci/38-GIT-PLUS.yaml, ci/39-PYPI.yaml, ci/39-PYPI-PLUS.yaml, ci/39-GIT.yaml, ci/39-GIT-PLUS.yaml]
+         environment-file: [ci/36-STABLE.yaml, ci/36-STABLE-PLUS.yaml, ci/36-DEV.yaml, ci/36-DEV-PLUS.yaml, ci/37-STABLE.yaml, ci/37-STABLE-PLUS.yaml, ci/37-DEV.yaml, ci/37-DEV-PLUS.yaml, ci/38-STABLE.yaml, ci/38-STABLE-PLUS.yaml, ci/38-DEV.yaml, ci/38-DEV-PLUS.yaml, ci/39-STABLE.yaml, ci/39-STABLE-PLUS.yaml, ci/39-DEV.yaml, ci/39-DEV-PLUS.yaml]
          exclude:
-           - environment-file: ci/36-PYPI-PLUS.yaml
+           - environment-file: ci/36-STABLE-PLUS.yaml
              os: windows-latest
-           - environment-file: ci/36-GIT-PLUS.yaml
+           - environment-file: ci/36-DEV-PLUS.yaml
              os: windows-latest
+           
      steps:
        - name: checkout repo
          uses: actions/checkout@v2
@@ -37,14 +38,14 @@
          run: |
            pip install pip
            pip install git+https://github.com/pysal/libpysal.git@master
-         if: matrix.os != 'windows-latest' && contains(matrix.environment-file, 'GIT')
+         if: matrix.os != 'windows-latest' && contains(matrix.environment-file, 'DEV')
       
        - name: install bleeding edge libpysal - powershell
          shell: powershell
          run: |
            pip install pip
            pip install git+https://github.com/pysal/libpysal.git@master
-         if: matrix.os == 'windows-latest' && contains(matrix.environment-file, 'GIT')
+         if: matrix.os == 'windows-latest' && contains(matrix.environment-file, 'DEV')
        
        - name: run tests - bash
          shell: bash -l {0}
