### 2021-04-25 fb363c1
add pip req / remove pip install
