+++ .github/workflows/unittests.yml
@@ -35,16 +35,12 @@
       
        - name: install bleeding edge libpysal - bash
          shell: bash -l {0}
-         run: |
-           pip install pip
-           pip install git+https://github.com/pysal/libpysal.git@master
+         run: pip install git+https://github.com/pysal/libpysal.git@master
          if: matrix.os != 'windows-latest' && contains(matrix.environment-file, 'DEV')
       
        - name: install bleeding edge libpysal - powershell
          shell: powershell
-         run: |
-           pip install pip
-           pip install git+https://github.com/pysal/libpysal.git@master
+         run: pip install git+https://github.com/pysal/libpysal.git@master
          if: matrix.os == 'windows-latest' && contains(matrix.environment-file, 'DEV')
        
        - name: run tests - bash
