+++ ci/38-DEV.yaml
@@ -7,6 +7,7 @@ dependencies:
   - esda
   - numpy
   - pandas>=1.0
+  - pip
   - rtree
   - scipy>=1.0
   - watermark
