### 2021-04-27 ccd5a20
rename and remove CI envs
