+++ ci/37-DEV.yaml
@@ -1,18 +1,0 @@
-name: test
-channels:
-  - conda-forge
-dependencies:
-  - python=3.7
-  # required
-  - esda
-  - numpy
-  - pandas>=1.0
-  - pip
-  - rtree
-  - scipy>=1.0
-  - watermark
-  # testing/formatting
-  - black
-  - codecov
-  - pytest
-  - pytest-cov
