+++ ci/36-STABLE.yaml
@@ -1,18 +1,0 @@
-name: test
-channels:
-  - conda-forge
-dependencies:
-  - python=3.6
-  # required
-  - libpysal
-  - esda
-  - numpy
-  - pandas>=1.0
-  - rtree
-  - scipy>=1.0
-  - watermark
-  # testing/formatting
-  - black
-  - codecov
-  - pytest
-  - pytest-cov
