### 2021-04-27 217ce1d
prune down tests in matrix
