+++ ci/38.yaml
@@ -5,6 +5,7 @@ dependencies:
   - python=3.8
   # required
   - esda
+  - libpysal
   - numpy
   - pandas>=1.0
   - pip
