+++ ci/39.yaml
@@ -5,6 +5,7 @@ dependencies:
   - python=3.9
   # required
   - esda
+  - libpysal
   - numpy
   - pandas>=1.0
   - pip
