+++ .github/workflows/unittests.yml
@@ -16,14 +16,23 @@
      strategy:
        matrix:
          os: [macos-latest, ubuntu-latest, windows-latest]
-         environment-file: [ci/36.yaml, ci/37.yaml, ci/38.yaml, ci/39.yaml, ci/39-BASE.yaml]
+         environment-file: [ci/36.yaml, ci/37.yaml, ci/38.yaml, ci/39.yaml]
          exclude:
            - environment-file: ci/36.yaml
              os: windows-latest
+           #- environment-file: ci/39-DEV.yaml
+           #  os: macos-latest
+           #- environment-file: ci/39-DEV.yaml
+           #  os: windows-latest
+           #- environment-file: ci/39-BASE.yaml
+           #  os: macos-latest
+           #- environment-file: ci/39-BASE.yaml
+           #  os: windows-latest
+         include:
            - environment-file: ci/39-BASE.yaml
-             os: macos-latest
-           - environment-file: ci/39-BASE.yaml
-             os: windows-latest
+             os: ubuntu-latest
+           - environment-file: ci/39-DEV.yaml
+             os: ubuntu-latest
            
      steps:
        - name: checkout repo
@@ -35,15 +44,10 @@
            environment-file: ${{ matrix.environment-file }}
            micromamba-version: 'latest'
       
-       - name: install bleeding edge libpysal - bash
+       - name: install bleeding edge libpysal (only Ubuntu / Python 3.9)
          shell: bash -l {0}
          run: pip install git+https://github.com/pysal/libpysal.git@master
-         if: matrix.os != 'windows-latest'
-      
-       - name: install bleeding edge libpysal - powershell
-         shell: powershell
-         run: pip install git+https://github.com/pysal/libpysal.git@master
-         if: matrix.os == 'windows-latest'
+         if: matrix.os == 'ubuntu-latest' && contains(matrix.environment-file, 'DEV')
        
        - name: run tests - bash
          shell: bash -l {0}
