+++ .github/workflows/unittests.yml
@@ -1,12 +1,14 @@
+ name Continuous Integration
+ 
  on:
    push:
      branches:
-     - '*'
+       - '*'
    pull_request:
      branches:
-     - '*'
+       - '*'
    schedule:
-     - cron: '59 23 * * *'
+       - cron: '59 23 * * *'
 
  jobs:
    unittests:
@@ -20,14 +22,6 @@
          exclude:
            - environment-file: ci/36.yaml
              os: windows-latest
-           #- environment-file: ci/39-DEV.yaml
-           #  os: macos-latest
-           #- environment-file: ci/39-DEV.yaml
-           #  os: windows-latest
-           #- environment-file: ci/39-BASE.yaml
-           #  os: macos-latest
-           #- environment-file: ci/39-BASE.yaml
-           #  os: windows-latest
          include:
            - environment-file: ci/39-BASE.yaml
              os: ubuntu-latest
