+++ .github/workflows/build_docs.yml
@@ -1,5 +1,5 @@
-
  name: Build Docs
+ 
  on:
    push:
      # Sequence of patterns matched against refs/tags
@@ -13,7 +13,7 @@
          required: false
  jobs:
    docs:
-     name: Build & Push Docs
+     name: build & push docs
      runs-on: ${{ matrix.os }}
      timeout-minutes: 90
      strategy:
@@ -24,32 +24,21 @@
      defaults:
        run:
          shell: bash -l {0}
+     
      steps:
-       - uses: actions/checkout@v2
-       - uses: actions/cache@v2
-         env:
-           CACHE_NUMBER: 0
+       - name: checkout repo
+         uses: actions/checkout@v2
+       
+       - name: setup micromamba
+         uses: mamba-org/provision-with-micromamba@main
          with:
-           path: ~/conda_pkgs_dir
-           key: ${{ matrix.os }}-conda-${{ env.CACHE_NUMBER }}-${{ hashFiles(matrix.environment-file) }}
-       - uses: conda-incubator/setup-miniconda@v2
-         with:
-            miniconda-version: 'latest'
-            mamba-version: '*'
-            channels: conda-forge
-            channel-priority: true
-            auto-update-conda: false
-            auto-activate-base: false
-            environment-file: ${{ matrix.environment-file }}
-            activate-environment: test
-            use-only-tar-bz2: true
-       - run: mamba info --all
-       - run: mamba list
-       - run: conda config --show-sources
-       - run: conda config --show
-       - name: Make Docs
+           environment-file: ${{ matrix.environment-file }}
+           micromamba-version: 'latest'
+       
+       - name: make docs
          run: cd docs; make html
-       - name: Commit Docs
+       
+       - name: commit docs
          run: |
            git clone https://github.com/ammaraskar/sphinx-action-test.git --branch gh-pages --single-branch gh-pages
            cp -r docs/_build/html/* gh-pages/
@@ -60,7 +49,9 @@
            git commit -m "Update documentation" -a || true
            # The above command will fail if no changes were present,
            # so we ignore the return code.
-       - uses: ad-m/github-push-action@master
+       
+       - name: push to gh-pages
+         uses: ad-m/github-push-action@master
          with:
             branch: gh-pages
             directory: gh-pages
