+++ .github/workflows/unittests.yml
@@ -1,4 +1,4 @@
- name Continuous Integration
+ name: Continuous Integration
  
  on:
    push:
