+++ .github/workflows/unittests.yml
@@ -54,7 +54,7 @@
          if: matrix.os == 'windows-latest'
        
        - name: codecov
-         uses: codecov/codecov-action@v1
+         uses: codecov/codecov-action@v1.5.0
          with:
            token: ${{ secrets.CODECOV_TOKEN }}
            file: ./coverage.xml
