### 2021-05-12 3796750
Bump actions/create-release from 1 to 1.1.4
Bumps [actions/create-release](https://github.com/actions/create-release) from 1 to 1.1.4.
- [Release notes](https://github.com/actions/create-release/releases)
- [Commits](https://github.com/actions/create-release/compare/v1...v1.1.4)

Signed-off-by: dependabot[bot] <support@github.com>
