+++ .github/workflows/release_and_publish.yml
@@ -44,7 +44,7 @@ jobs:
       
       - name: create release
         id: create_release
-        uses: actions/create-release@v1
+        uses: actions/create-release@v1.1.4
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # token is provided by GHA, DO NOT create
         with:
