+++ .github/workflows/release_and_publish.yml
@@ -62,7 +62,7 @@ jobs:
       
       - name: upload release asset
         id: upload-release-asset
-        uses: actions/upload-release-asset@v1
+        uses: actions/upload-release-asset@v1.0.2
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         with:
