+++ .github/workflows/release_and_publish.yml
@@ -20,7 +20,7 @@ jobs:
         uses: actions/checkout@v2
       
       - name: set up python
-        uses: actions/setup-python@v2
+        uses: actions/setup-python@v2.2.2
         with:
           python-version: '3.x'
       
