+++ .github/workflows/unittests.yml
@@ -30,7 +30,7 @@
            
      steps:
        - name: checkout repo
-         uses: actions/checkout@v2
+         uses: actions/checkout@v2.3.4
        
        - name: setup micromamba
          uses: mamba-org/provision-with-micromamba@main
