+++ .github/workflows/release_and_publish.yml
@@ -17,7 +17,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: checkout repo
-        uses: actions/checkout@v2
+        uses: actions/checkout@v2.3.4
       
       - name: set up python
         uses: actions/setup-python@v2
