### 2021-05-31 pysal/spaghetti#575 @martibosch
Hello again @jGaboardi,

in my opinion, the current statement of need should be shortened so that it focuses more on what the target audience can do with spaghetti that cannot be done with other existing tools (e.g., read and analyze network data within the PySAL ecosystem) rather than reviewing the timeline of how PySAL started and how spaghetti became a separate component. This historical perspective could be somewhere in the documentation site, but I believe that the JOSS paper should be as brief as possible and focus on the functionalities and what users can get from the library from a more practical standpoint.

Best,
Martí
