### 2021-05-31 #2826 @martibosch
Hello all. First of all my apologies for my delay - these last month with the end of my thesis have been quite busy.

all my comments have been addressed and I would just like to suggest a final iteration to make the current statement of need as synthetic as possible - to that end, I have reopened [#575](https://github.com/pysal/spaghetti/issues/575). After this, I believe the manuscript can go to publication.

Best,
Martí
