### 2021-06-02 pysal/spaghetti#575 @jGaboardi
@martibosch I have trimmed down the Statement of Need section by adding the topic sentence of the second paragraph into the first paragraph and removing the second paragraph altogether.
