+++ spaghetti/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.5.7"
+__version__ = "1.5.8"
 
 """
 # `spaghetti` --- Spatial Graphs: Networks, Topology, & Inference
