### 2021-06-03 #2826 @jGaboardi
@arfon I have cut a fresh release and Zenodo archive with the following info. Let me know if I did this correctly.
| title | spaghetti: spatial network analysis in PySAL |
|  --- | --- |
| version | 1.5.8 |
| release | https://github.com/pysal/spaghetti/releases/tag/v1.5.8 |
|  doi | https://zenodo.org/record/4898079 |

@slumnitz @sjsrey looks like things are wrapping up here!
