### 2021-06-03 #2826 @arfon
@jGaboardi – please merge this PR which makes a small number of formatting fixes and removes a duplicate sentence https://github.com/pysal/spaghetti/pull/626.
