### 2021-06-03 #2826 @jGaboardi
@arfon pysal/spaghetti#626 is now merged.
