### 2021-06-04 #2826 @arfon
@martibosch, @usethedata – many thanks for your reviews here and to @usethedata for editing too! JOSS relies upon the volunteer efforts of people like you and we simply wouldn't be able to do this without you ✨

On a personal note, thanks to @usethedata who is stepping down from the JOSS editorial team. Thanks Bruce for all of your help making JOSS the success it is 💖
